% Generated by roxygen2 (4.0.0): do not edit by hand
\name{add_build_ignore}
\alias{add_build_ignore}
\title{Add a file to \code{.Rbuildignore}}
\usage{
add_build_ignore(pkg = ".", file, escape = TRUE)
}
\arguments{
\item{pkg}{package description, can be path or package name.  See
\code{\link{as.package}} for more information}

\item{file}{Name of file.}

\item{escape}{If \code{TRUE}, the default, will escape \code{.} to
\code{\\.} and surround with \code{^} and \code{$}.}
}
\value{
Nothing, called for its side effect.
}
\description{
\code{.Rbuildignore} has a regular expression on each line, but it's
usually easier to work with specific file names. By default, will (crudely)
turn a filename into a regular expression that will only match that
path. Repeated entries will be silently removed.
}
\keyword{internal}

