% Generated by roxygen2 (4.0.0): do not edit by hand
\name{add_travis}
\alias{add_travis}
\title{Add basic travis template to a package}
\usage{
add_travis(pkg = ".")
}
\arguments{
\item{pkg}{package description, can be path or package name.  See
\code{\link{as.package}} for more information}
}
\description{
Also adds \code{.travis.yml} to \code{.Rbuildignore} so it isn't included
in the built package
}

