\name{bisect_load_and_test}
\alias{bisect_load_and_test}
\title{Loads a source package from a directory and runs a test function.}
\usage{
  bisect_load_and_test(fun, pkgdir = ".",
    on_load_error = NA, on_run_error = NA)
}
\description{
  Loads a source package from a directory and runs a test
  function.
}
\seealso{
  \code{\link{bisect_runtest}}

  \code{\link{bisect_load_all}}

  \code{\link{bisect_return_interactive}}
}

