\name{bisect_return_interactive}
\alias{bisect_return_interactive}
\title{Prompt the user for an interactive good/bad/skip response and return
the appropriate value (to be passed to \code{bisect_runtest}).}
\usage{
  bisect_return_interactive()
}
\description{
  Prompt the user for an interactive good/bad/skip response
  and return the appropriate value (to be passed to
  \code{bisect_runtest}).
}
\seealso{
  \code{\link{bisect_runtest}}

  \code{\link{bisect_load_all}}

  \code{\link{bisect_install}}
}

