% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/install-bitbucket.r
\name{bitbucket_pull}
\alias{bitbucket_pull}
\title{Bitbucket references}
\usage{
bitbucket_pull(pull)
}
\arguments{
\item{pull}{The pull request to install}
}
\description{
Use as \code{ref} parameter to \code{\link{install_bitbucket}}.
Allows installing a specific pull request.
}
\seealso{
\code{\link{install_bitbucket}}
}

