% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build-manual.R
\name{build_manual}
\alias{build_manual}
\title{Create package pdf manual}
\usage{
build_manual(pkg = ".", path = NULL)
}
\arguments{
\item{pkg}{package description, can be path or package name.  See
\code{\link[=as.package]{as.package()}} for more information.}

\item{path}{path in which to produce package manual.
If \code{NULL}, defaults to the parent directory of the package.}
}
\description{
Create package pdf manual
}
\seealso{
\code{\link[=Rd2pdf]{Rd2pdf()}}
}
