% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build-readme.R
\name{build_md}
\alias{build_md}
\alias{build_rmd}
\alias{build_readme}
\title{Build a qmd or Rmarkdown files package}
\usage{
build_md(files, path = ".", output_options = list(), ..., quiet = TRUE)

build_rmd(...)

build_readme(path = ".", quiet = TRUE, ...)
}
\arguments{
\item{files}{The R Markdown or Quarto files to be rendered.}

\item{path}{path to the package to build the readme.}

\item{output_options}{A list of options passed to \code{\link[rmarkdown:render]{rmarkdown::render()}} for
\code{.Rmd} inputs. Ignored for Quarto \code{.qmd} inputs.}

\item{...}{additional arguments passed to \code{\link[rmarkdown:render]{rmarkdown::render()}} for \code{.Rmd}
inputs or to \code{\link[quarto:quarto_render]{quarto::quarto_render()}} for \code{.qmd} inputs. Arguments are shared.}

\item{quiet}{If \code{TRUE}, suppress output.}
}
\description{
\code{build_md()} is a wrapper around \code{\link[rmarkdown:render]{rmarkdown::render()}} and
\code{\link[quarto:quarto_render]{quarto::quarto_render()}} that first installs a temporary copy of the
package, and then renders each \code{.Rmd} or \code{.qmd} in a clean R session.
\code{build_readme()} locates your \code{README.Rmd} or \code{README.qmd} and builds it into
a \code{README.md}.
}
