% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build-readme.R
\name{build_rmd}
\alias{build_rmd}
\title{Build Rmarkdown files}
\usage{
build_rmd(files, path = ".", output_options = list(), ..., quiet = TRUE)
}
\arguments{
\item{files}{The Rmarkdown files to be rendered.}

\item{path}{path to the package to build the readme.}

\item{output_options}{List of output options that can override the options
specified in metadata (e.g. could be used to force \code{self_contained} or
\code{mathjax = "local"}). Note that this is only valid when the output
format is read from metadata (i.e. not a custom format object passed to
\code{output_format}).}

\item{...}{additional arguments passed to \code{\link[rmarkdown:render]{rmarkdown::render()}}}

\item{quiet}{If \code{TRUE}, suppress output.}
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#deprecated}{\figure{lifecycle-deprecated.svg}{options: alt='[Deprecated]'}}}{\strong{[Deprecated]}}

\code{build_rmd()} is deprecated, as it is a low-level helper for internal use. To
render your package's \code{README.qmd} or \code{README.Rmd}, use \code{\link[=build_readme]{build_readme()}}. To
preview a vignette or article, use functions like \code{\link[pkgdown:build_site]{pkgdown::build_site()}} or
\code{\link[pkgdown:build_article]{pkgdown::build_article()}}.
}
\keyword{internal}
