% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build-site.R
\name{build_site}
\alias{build_site}
\title{Execute \pkg{pkgdown} build_site in a package}
\usage{
build_site(path = ".", quiet = TRUE, ...)
}
\arguments{
\item{path}{path to the package to build the static HTML.}

\item{quiet}{If \code{TRUE}, suppress output.}

\item{...}{additional arguments passed to \code{\link[pkgdown:build_site]{pkgdown::build_site()}}}
}
\description{
\code{build_site()} is a shortcut for \code{\link[pkgdown:build_site]{pkgdown::build_site()}}, it generates the
static HTML documentation.
}
