\name{bundle}
\alias{bundle}
\title{Bundles a package's dependencies for development workflows.}
\usage{
  bundle(pkg = NULL)
}
\arguments{
  \item{pkg}{package description, can be path or package
  name.  See \code{\link{as.package}} for more information}
}
\description{
  Dependencies are installed into the package's bundle
  library. See bundle_library.
}

