\name{changed_files}
\alias{changed_files}
\title{Given vector of paths, return only those paths that have changed since the
last invocation.}
\usage{
  changed_files(paths)
}
\description{
  Given vector of paths, return only those paths that have
  changed since the last invocation.
}
\keyword{internal}

