% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/check-coverage.r
\name{check_coverage}
\alias{check_coverage}
\title{Check test coverage for a package.}
\usage{
check_coverage(pkg = ".", ...)
}
\arguments{
\item{pkg}{package description, can be path or package name. See
\code{\link{as.package}} for more information}

\item{...}{additional arguments passed to \code{\link[covr]{package_coverage}}}
}
\description{
This function runs \code{\link[covr]{package_coverage}} to track and
calculate test coverage for a package.
}
\seealso{
\code{\link[covr]{package_coverage}}
}

