\name{check_dep_version}
\alias{check_dep_version}
\title{Check that the version of an imported package satisfies the requirements}
\usage{
check_dep_version(dep_name, dep_ver = NA, dep_compare = NA)
}
\arguments{
  \item{dep_name}{The name of the package with objects to
  import}

  \item{dep_ver}{The version of the package}

  \item{dep_compare}{The comparison operator to use to
  check the version}
}
\description{
Check that the version of an imported package satisfies the
requirements
}
\keyword{internal}

