% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check-doc.R
\name{check_doc_fields}
\alias{check_doc_fields}
\title{Check for missing documentation fields}
\usage{
check_doc_fields(pkg = ".", fields = c("value", "examples"))
}
\arguments{
\item{pkg}{The package to use, can be a file path to the package or a
package object.  See \code{\link[=as.package]{as.package()}} for more information.}

\item{fields}{A character vector of Rd field names to check for.}
}
\value{
A named list of character vectors, one for each field, containing
the names of Rd files missing that field. Returned invisibly.
}
\description{
Checks all Rd files in \verb{man/} and looks for any that have a \verb{\\usage} section
(i.e. a function) but that \emph{don't} have \verb{\\value} and \verb{\\examples} sections.
These missing fields are flagged by CRAN on initial submission.
}
\examples{
\dontrun{
check_doc_fields(".")
}
}
