% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check-mac.R
\name{check_mac}
\alias{check_mac}
\alias{check_mac_devel}
\alias{check_mac_release}
\title{Check a package on macOS}
\usage{
check_mac_devel(
  pkg = ".",
  dep_pkgs = character(),
  args = NULL,
  manual = TRUE,
  quiet = FALSE,
  ...
)

check_mac_release(
  pkg = ".",
  dep_pkgs = character(),
  args = NULL,
  manual = TRUE,
  quiet = FALSE,
  ...
)
}
\arguments{
\item{pkg}{The package to use, can be a file path to the package or a
package object.  See \code{\link[=as.package]{as.package()}} for more information.}

\item{dep_pkgs}{Additional custom dependencies to install prior to checking
the package.}

\item{args}{An optional character vector of additional command
line arguments to be passed to \verb{R CMD build} if \code{binary = FALSE},
or \verb{R CMD install} if \code{binary = TRUE}.}

\item{manual}{Should the manual be built?}

\item{quiet}{If \code{TRUE}, suppresses output.}

\item{...}{Additional arguments passed to \code{\link[pkgbuild:build]{pkgbuild::build()}}.}
}
\value{
The url with the check results (invisibly)
}
\description{
This function first bundles a source package, then uploads it to
\url{https://mac.r-project.org/macbuilder/submit.html}. This function returns a
link to the page where the check results will appear.
}
\section{Functions}{
\itemize{
\item \code{check_mac_devel()}: Check package on the development version of R.

\item \code{check_mac_release()}: Check package on the released version of R.

}}
\seealso{
Other build functions: 
\code{\link{check_win}()}
}
\concept{build functions}
