% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check-doc.R
\name{check_man}
\alias{check_man}
\title{Check documentation, as \verb{R CMD check} does}
\usage{
check_man(pkg = ".")
}
\arguments{
\item{pkg}{The package to use, can be a file path to the package or a
package object.  See \code{\link[=as.package]{as.package()}} for more information.}
}
\value{
Nothing. This function is called purely for it's side effects: if
no errors there will be no output.
}
\description{
This function attempts to run the documentation related checks in the
same way that \verb{R CMD check} does. Unfortunately it can't run them
all because some tests require the package to be loaded, and the way
they attempt to load the code conflicts with how devtools does it.
}
\examples{
\dontrun{
check_man("mypkg")
}
}
