% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/complete.R
\name{complete}
\alias{complete}
\title{Complete development session}
\usage{
complete(pkg = ".")
}
\arguments{
\item{pkg}{package description, can be path or package name.  See
\code{\link{as.package}} for more information}
}
\description{
helps complete a session by checking if NEWS(.md), version, and date are up-to-date
}

