\name{create_package_doc}
\alias{create_package_doc}
\title{Creates package-level documentation, where the developer can add
documentation for data, vignettes, etc.}
\usage{
  create_package_doc(path, name)
}
\arguments{
  \item{path}{path to the package}

  \item{name}{of the package}
}
\description{
  Creates package-level documentation, where the developer
  can add documentation for data, vignettes, etc.
}

