% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/session-info.r
\name{dev_packages}
\alias{dev_packages}
\title{Return a vector of names of packages loaded by devtools}
\usage{
dev_packages()
}
\description{
Return a vector of names of packages loaded by devtools
}
\keyword{internal}

