% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sitrep.R
\name{dev_sitrep}
\alias{dev_sitrep}
\title{Report package development situation}
\usage{
dev_sitrep(pkg = ".", debug = FALSE)
}
\arguments{
\item{pkg}{The package to use, can be a file path to the package or a
package object.  See \code{\link[=as.package]{as.package()}} for more information.}

\item{debug}{If \code{TRUE}, will print out extra information useful for
debugging. If \code{FALSE}, it will use result cached from a previous run.}
}
\value{
A named list, with S3 class \code{dev_sitrep} (for printing purposes).
}
\description{
Call this function if things seem weird and you're not sure
what's wrong or how to fix it. It reports:
\itemize{
\item If R is up to date.
\item If RStudio is up to date.
\item If compiler build tools are installed and available for use.
\item If devtools and its dependencies are up to date.
\item If the package's dependencies are up to date.
}
}
\examples{
\dontrun{
dev_sitrep()
}
}
