% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/test.r
\name{devtest}
\alias{devtest}
\title{Return the path to one of the packages in the devtools test dir}
\usage{
devtest(package)
}
\arguments{
\item{package}{Name of the test package.}
}
\description{
Devtools comes with some simple packages for testing. This function
returns the path to them.
}
\examples{
if (has_tests()) {
devtest("testData")
}
}
\keyword{internal}

