% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test.R
\name{devtools-defunct}
\alias{devtools-defunct}
\alias{test_file}
\alias{test_coverage_file}
\title{Defunct functions}
\usage{
test_file(file = find_active_file(), ...)

test_coverage_file(file = find_active_file(), ...)
}
\description{
These functions are defunct and will be removed in a future version of
devtools.
}
\keyword{internal}
