% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/infrastructure-git.R, R/infrastructure.R,
%   R/revdep-email.R, R/revdep-summarise.R, R/revdep.R, R/zzz.r
\name{use_git}
\alias{use_git}
\alias{use_github}
\alias{use_git_hook}
\alias{use_github_links}
\alias{use_testthat}
\alias{use_test}
\alias{use_rstudio}
\alias{use_vignette}
\alias{use_rcpp}
\alias{use_travis}
\alias{use_coverage}
\alias{use_appveyor}
\alias{use_package_doc}
\alias{use_package}
\alias{use_data}
\alias{use_data_raw}
\alias{use_build_ignore}
\alias{use_readme_rmd}
\alias{use_readme_md}
\alias{use_news_md}
\alias{use_revdep}
\alias{use_cran_comments}
\alias{use_code_of_conduct}
\alias{use_cran_badge}
\alias{use_mit_license}
\alias{use_gpl3_license}
\alias{use_dev_version}
\alias{revdep_email}
\alias{revdep_check_save_summary}
\alias{revdep_check_print_problems}
\alias{revdep_check}
\alias{revdep_check_resume}
\alias{revdep_check_reset}
\alias{devtools-deprecated}
\title{Deprecated Functions}
\usage{
use_git(..., pkg = ".")

use_github(..., pkg = ".")

use_git_hook(..., pkg = ".")

use_github_links(..., pkg = ".")

use_testthat(..., pkg = ".")

use_test(..., pkg = ".")

use_rstudio(..., pkg = ".")

use_vignette(..., pkg = ".")

use_rcpp(..., pkg = ".")

use_travis(..., pkg = ".")

use_coverage(..., pkg = ".")

use_appveyor(..., pkg = ".")

use_package_doc(..., pkg = ".")

use_package(..., pkg = ".")

use_data(..., pkg = ".")

use_data_raw(..., pkg = ".")

use_build_ignore(..., pkg = ".")

use_readme_rmd(..., pkg = ".")

use_readme_md(..., pkg = ".")

use_news_md(..., pkg = ".")

use_revdep(..., pkg = ".")

use_cran_comments(..., pkg = ".")

use_code_of_conduct(..., pkg = ".")

use_cran_badge(..., pkg = ".")

use_mit_license(..., pkg = ".")

use_gpl3_license(..., pkg = ".")

use_dev_version(..., pkg = ".")

revdep_email(pkg = ".", date, version, author = getOption("devtools.name"),
  draft = TRUE, unsent = NULL, template = "revdep/email.md",
  only_problems = TRUE)

revdep_check_save_summary(pkg = ".")

revdep_check_print_problems(pkg = ".")

revdep_check(pkg = ".", recursive = FALSE, ignore = NULL,
  dependencies = c("Depends", "Imports", "Suggests", "LinkingTo"),
  skip = character(), libpath = getOption("devtools.revdep.libpath"),
  srcpath = libpath, bioconductor = FALSE, type = getOption("pkgType"),
  threads = getOption("Ncpus", 1), env_vars = NULL, check_dir = NULL,
  install_dir = NULL, quiet_check = TRUE)

revdep_check_resume(pkg = ".", ...)

revdep_check_reset(pkg = ".")
}
\description{
These functions are Deprecated in this release of devtools, they will be
marked as Defunct and removed in a future version.
}
\details{
Instead of the use_xyz functions from devtools use \link[pkg:usethis]{use_testthat}.

Instead of the revdep functions in devtools a better alternative is
\link[revdepcheck]{revdep_check}.
}
\keyword{internal}
