% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/devtools-package.R
\docType{package}
\name{devtools-package}
\alias{devtools}
\alias{devtools-package}
\title{devtools: Tools to Make Developing R Packages Easier}
\description{
\if{html}{\figure{logo.png}{options: style='float: right' alt='logo' width='120'}}

Collection of package development tools.
}
\section{Package options}{


Devtools uses the following \code{\link[=options]{options()}} to configure behaviour:
\itemize{
\item \code{devtools.install.args}: a string giving extra arguments passed to
\verb{R CMD install} by \code{\link[=install]{install()}}.
\item \code{devtools.path}: Deprecated. Path used by the now-deprecated \code{\link[=dev_mode]{dev_mode()}}
function.
}
}

\seealso{
Useful links:
\itemize{
  \item \url{https://devtools.r-lib.org/}
  \item \url{https://github.com/r-lib/devtools}
  \item Report bugs at \url{https://github.com/r-lib/devtools/issues}
}

}
\author{
\strong{Maintainer}: Jennifer Bryan \email{jenny@posit.co} (\href{https://orcid.org/0000-0002-6983-2759}{ORCID})

Authors:
\itemize{
  \item Jennifer Bryan \email{jenny@posit.co} (\href{https://orcid.org/0000-0002-6983-2759}{ORCID})
  \item Hadley Wickham
  \item Jim Hester
  \item Winston Chang
}

Other contributors:
\itemize{
  \item Posit Software, PBC (\href{https://ror.org/03wc8by49}{ROR}) [copyright holder, funder]
}

}
\keyword{internal}
