% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/document.R
\name{document}
\alias{document}
\title{Use roxygen to document a package}
\usage{
document(pkg = ".", roclets = NULL, quiet = FALSE)
}
\arguments{
\item{pkg}{The package to use, can be a file path to the package or a
package object.  See \code{\link[=as.package]{as.package()}} for more information.}

\item{roclets}{Character vector of \link[roxygen2:roclet]{roclets} to use.

The default, \code{NULL}, uses the roxygen \code{roclets} option,
which defaults to \code{c("collate", "namespace", "rd")}. This will update
(if needed) the \code{Collate} field with \code{\link[roxygen2:update_collate]{update_collate()}},
produce the \code{NAMESPACE} file with \code{\link[roxygen2:namespace_roclet]{namespace_roclet()}}, and
produce the Rd files with \code{\link[roxygen2:rd_roclet]{rd_roclet()}}.

(Note that \code{update_collate()} is not technically a roclet but is still
controlled with this argument for historical reasons.)}

\item{quiet}{if \code{TRUE} suppresses output from this function.}
}
\description{
This function is a wrapper for the \code{\link[roxygen2:roxygenize]{roxygen2::roxygenize()}}
function from the roxygen2 package. See the documentation and vignettes of
that package to learn how to use roxygen.
}
\seealso{
\code{\link[roxygen2:roxygenize]{roxygen2::roxygenize()}},
\code{browseVignettes("roxygen2")}
}
