% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/doctor.R
\name{dr_devtools}
\alias{dr_devtools}
\title{Diagnose potential devtools issues}
\usage{
dr_devtools()
}
\description{
This checks to make sure you're using the latest release of R,
the released version of RStudio (if you're using it as your gui),
and the latest version of devtools and its dependencies.
}
\examples{
\donttest{
dr_devtools()
}
}
\seealso{
Other doctors: \code{\link{dr_github}}
}

