% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/topic-index.r
\name{find_topic}
\alias{find_topic}
\title{Find the rd file that documents a topic.}
\usage{
find_topic(topic)
}
\arguments{
\item{topic}{The topic, a string.}
}
\value{
A named string. The values gives the path to file; the name gives
  the path to package.
}
\description{
Only packages loaded by devtools are searched.
}
\examples{
find_topic("help")
}
\keyword{internal}

