\name{functions_with_arg}
\alias{functions_with_arg}
\title{Find all functions with specified argument.}
\usage{functions_with_arg(arg, pos)}

\description{
  Find all functions with specified argument.
}
\keyword{programming}
\examples{\dontrun{
functions_with_arg("drop", "package:base")
functions_with_arg("na.rm", "package:base")
}}
