\name{git_add}
\alias{git_add}
\title{Stage files for commit}
\usage{
  git_add(file)
}
\arguments{
  \item{file}{a vector of files to add to the repository.}
}
\description{
  Stage files for commit
}
\seealso{
  Other git: \code{\link{git_add_new}},
  \code{\link{git_commit}}, \code{\link{git_commit_all}},
  \code{\link{git_status}}, \code{\link{git_tag}},
  \code{\link{has_git}}, \code{\link{win_has_git}}
}

