\name{git_add_new}
\alias{git_add_new}
\title{Stage all files for a commit.}
\usage{
  git_add_new(path = ".")
}
\arguments{
  \item{path}{The path to the files.}
}
\description{
  Stage all files for a commit.
}
\seealso{
  Other git: \code{\link{git_add}},
  \code{\link{git_commit}}, \code{\link{git_commit_all}},
  \code{\link{git_status}}, \code{\link{git_tag}},
  \code{\link{has_git}}, \code{\link{win_has_git}}
}

