% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check-git.R
\name{git_checks}
\alias{git_checks}
\title{Git checks}
\usage{
git_checks(pkg = ".")
}
\arguments{
\item{pkg}{The package to use, can be a file path to the package or a
package object.  See \code{\link[=as.package]{as.package()}} for more information.}
}
\description{
This function performs Git checks checks prior to release. It is called
automatically by \code{\link[=release]{release()}}.
}
\keyword{internal}
