% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/git.R
\name{git_wd_clean}
\alias{git_wd_clean}
\title{git_wd_clean}
\usage{
git_wd_clean(path = ".", level = "warn")
}
\arguments{
\item{path}{the path to the package directory}

\item{level}{how should this behave if wd is not clean (see details)}
}
\value{
nothing. Used for it's side effect
}
\description{
checks that the working directory (wd) is clean, i.e. the last SHA in commits
represents the current package structure.
}
\details{
This function controls behavior for calling functions on a \emph{git repo}.
The \code{level} parameter decides how fussy the function is about seeing
a git repo that has \emph{uncommitted} changes.

\describe{
  \item{none}{The function doesn't care at all}
  \item{warn}{A warning that the wd is not clean is displayed}
  \item{die}{An error is thrown, stopping all further functions}
 }
}

