% Generated by roxygen2 (4.0.1): do not edit by hand
\name{github_pull}
\alias{github_pull}
\title{Install a specific pull request from GitHub}
\usage{
github_pull(pull)
}
\arguments{
\item{pull}{The pull request to install}
}
\description{
Use as \code{ref} parameter to \code{\link{install_github}}.
}
\seealso{
\code{\link{install_github}}
}

