% Generated by roxygen2 (4.0.1): do not edit by hand
\name{github_ref}
\alias{github_ref}
\title{Resolve a token to a GitHub reference}
\usage{
github_ref(x, param)
}
\arguments{
\item{x}{Reference token}

\item{param}{A named list of GitHub parameters}
}
\description{
A generic function, for internal use only.
}
\keyword{internal}

