% Generated by roxygen2 (4.0.1): do not edit by hand
\name{github_release}
\alias{github_release}
\title{Install the latest release from GitHub}
\usage{
github_release()
}
\description{
Use as \code{ref} parameter to \code{\link{install_github}}.
}
\seealso{
\code{\link{install_github}}, \code{\link{github_pull}}
}

