% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/install-gitlab.r
\name{gitlab_pat}
\alias{gitlab_pat}
\title{Retrieve Gitlab personal access token.}
\usage{
gitlab_pat(quiet = TRUE)
}
\description{
A github personal access token
Looks in env var \code{GITLAB_PAT}
}
\keyword{internal}
