\name{has_git}
\alias{has_git}
\alias{win_has_git}
\title{Check for the presence of git on the system.}
\usage{
  has_git()

  win_has_git()
}
\description{
  Check to see if git is present. \code{win_has_git} will
  automatically add git to the path when called and if it
  finds git in the standard installation folder.
}
\seealso{
  Other git: \code{\link{git_add}},
  \code{\link{git_add_new}}, \code{\link{git_commit}},
  \code{\link{git_commit_all}}, \code{\link{git_status}},
  \code{\link{git_tag}}
}

