% Generated by roxygen2 (4.0.0): do not edit by hand
\name{help}
\alias{help}
\title{Drop-in replacement for help function}
\usage{
help(topic, package = NULL, ...)
}
\arguments{
\item{...}{Additional arguments to pass to \code{\link[utils]{help}}.}

\item{topic}{usually, a \link{name} or character string specifying the
   topic for which help is sought.  A character string (enclosed in
   explicit single or double quotes) is always taken as naming a topic.

   If the value of \code{topic} is a length-one
   character vector the topic is taken to be the value of the only
   element.  Otherwise \code{topic} must be a name or a \link{reserved}
   word (if syntactically valid) or character string.

   See \sQuote{Details} for what happens if this is omitted.
 }

\item{package}{a name or character vector giving the packages to look
   into for documentation, or \code{NULL}.  By default, all packages in
   the search path are used.  To avoid a name being deparsed use e.g.
   \code{(pkg_ref)} (see the examples).}
}
\description{
This function is a replacement for \code{\link[utils]{help}}. If
\code{package} is not specified, it will search for help in devtools-loaded
packages first, then in regular packages.
}
\details{
If \code{package} is specified, then it will search for help in
devtools-loaded packages or regular packages, as appropriate.
}

