\name{imports_env}
\alias{imports_env}
\title{Package imports environment
Contains objects imported from other packages. Is the parent of the
package namespace environment, and is a child of <namespace:base>,
which is a child of R_GlobalEnv,}
\usage{
  imports_env(pkg = NULL)
}
\description{
  Package imports environment Contains objects imported
  from other packages. Is the parent of the package
  namespace environment, and is a child of
  <namespace:base>, which is a child of R_GlobalEnv,
}

