% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/infrastructure.R
\name{infrastructure}
\alias{infrastructure}
\title{Add useful infrastructure to a package.}
\arguments{
\item{pkg}{package description, can be path or package name. See
\code{\link{as.package}} for more information.}
}
\description{
Add useful infrastructure to a package.
}
