% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/remotes.R
\name{install-deprecated}
\alias{install-deprecated}
\alias{install_bioc}
\alias{install_bitbucket}
\alias{install_cran}
\alias{install_dev}
\alias{install_git}
\alias{install_github}
\alias{install_gitlab}
\alias{install_local}
\alias{install_svn}
\alias{install_url}
\alias{install_version}
\alias{update_packages}
\alias{dev_package_deps}
\alias{github_pull}
\alias{github_release}
\title{Deprecated package installation functions}
\usage{
install_bioc(...)

install_bitbucket(...)

install_cran(...)

install_dev(...)

install_git(...)

install_github(...)

install_gitlab(...)

install_local(...)

install_svn(...)

install_url(...)

install_version(...)

update_packages(...)

dev_package_deps(...)

github_pull(...)

github_release(...)
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#deprecated}{\figure{lifecycle-deprecated.svg}{options: alt='[Deprecated]'}}}{\strong{[Deprecated]}}

These functions have been deprecated in favor of \href{https://pak.r-lib.org/}{pak},
as that is what we now recommend for package installation. There are a few
functions which have no pak equivalent, where you can instead call the
old remotes functions directly.
\subsection{Migration guide}{\tabular{ll}{
   devtools function \tab Replacement \cr
   \code{install_bioc("pkg")} \tab \code{pak::pak("bioc::pkg")} \cr
   \code{install_bitbucket("user/repo")} \tab \code{remotes::install_bitbucket("user/repo")} \cr
   \code{install_cran("pkg")} \tab \code{pak::pak("pkg")} \cr
   \code{install_dev("pkg")} \tab \code{remotes::install_dev("pkg")} \cr
   \code{install_git("url")} \tab \code{pak::pak("git::url")} \cr
   \code{install_github("user/repo")} \tab \code{pak::pak("user/repo")} \cr
   \code{install_gitlab("user/repo")} \tab \code{pak::pak("gitlab::user/repo")} \cr
   \code{install_local("path")} \tab \code{pak::pak("local::path")} \cr
   \code{install_svn("url")} \tab \code{remotes::install_svn("url")} \cr
   \code{install_url("url")} \tab \code{pak::pak("url::url")} \cr
   \code{install_version("pkg", "1.0.0")} \tab \code{pak::pak("pkg@1.0.0")} \cr
   \code{update_packages("pkg")} \tab \code{pak::pak("pkg")} \cr
   \code{dev_package_deps()} \tab \code{pak::local_dev_deps()} \cr
   \code{github_pull("123")} \tab \code{remotes::github_pull("123")} \cr
   \code{github_release()} \tab \code{remotes::github_release()} \cr
}

}
}
\keyword{internal}
