% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/install-bitbucket.r
\name{install_bitbucket}
\alias{install_bitbucket}
\title{Install a package directly from bitbucket}
\usage{
install_bitbucket(repo, username = NULL, ref = "master", subdir = NULL,
  auth_token = bitbucket_pat(), host = "https://api.bitbucket.org",
  quiet = FALSE, ...)
}
\arguments{
\item{repo}{Repository address in the format
\code{username/repo[/subdir][@ref|#pull]}. Alternatively, you can
specify \code{subdir} and/or \code{ref} using the respective parameters
(see below); if both are specified, the values in \code{repo} take
precedence.}

\item{username}{User name. Deprecated: please include username in the
\code{repo}}

\item{ref}{Desired git reference. Could be a commit, tag, or branch
name, or a call to \code{\link{bitbucket_pull}}. Defaults to \code{"master"}.}

\item{subdir}{subdirectory within repo that contains the R package.}

\item{auth_token}{see \code{Details} section. PATs can be created at
\url{https://bitbucket.org/account/admin/app-passwords}}

\item{host}{GitHub API host to use. Override with your GitHub enterprise
hostname, for example, \code{"github.hostname.com/api/v3"}.}

\item{quiet}{if \code{TRUE} suppresses output from this function.}

\item{...}{Other arguments passed on to \code{\link{install}}.}
}
\description{
This function is vectorised so you can install multiple packages in
a single command.
}
\details{
To install from a private repo, or more generally, access the Bitbucket API
with your own credentials, you will need to get an access token. You can
create an access token following the instructions found in the
\href{https://confluence.atlassian.com/bitbucket/app-passwords-828781300.html}{Bitbucket App Passwords documentation}.
This PAT requires read-only access to your repositories and pull requests.
}
\examples{
\dontrun{
install_bitbucket("dannavarro/lsr-package")
}
}
\seealso{
Other package installation: \code{\link{install_bioc}},
  \code{\link{install_cran}}, \code{\link{install_github}},
  \code{\link{install_git}}, \code{\link{install_svn}},
  \code{\link{install_url}}, \code{\link{install_version}},
  \code{\link{install}}, \code{\link{uninstall}}
}

