% Generated by roxygen2 (4.0.1): do not edit by hand
\name{install_github_enterprise}
\alias{install_github_enterprise}
\title{Convenience wrapper for \code{\link{install_github}}.}
\usage{
install_github_enterprise(repo, username = getOption("github.user"),
  ref = "master", pull = NULL, subdir = NULL, branch = NULL,
  auth_user = NULL, password = NULL, auth_token = github_pat(),
  github_url = devtools_git_enterprise(), ..., dependencies = TRUE)
}
\description{
This function allows you to install a package built
with \code{devtools} from a repo with a custom URL.
}
\examples{
\dontrun{
# To install from a private repo, use auth_token as described
# at ?install_github and either set the github_url parameter
# or let the ?devtools_git_enterprise helper retrieve it from
# the GITHUB_URL environment variable if it is set.
# Best practice is the latter.
install_github_enterprise("username/packagename", github_url = "https://github.scm.xyz.com")

}
}
\seealso{
Other package installation: \code{\link{install_bitbucket}};
  \code{\link{install_github}};
  \code{\link{install_gitorious}};
  \code{\link{install_git}}; \code{\link{install_svn}};
  \code{\link{install_url}}; \code{\link{install_version}};
  \code{\link{install}}
}

