% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/install-gitlab.r
\name{install_gitlab}
\alias{install_gitlab}
\title{Attempts to install a package directly from GitLab}
\usage{
install_gitlab(repo, auth_token = gitlab_pat(),
  host = "https://www.gitlab.com", quiet = FALSE, ...)
}
\arguments{
\item{repo}{Repository address in the format
\code{username/repo[/subdir][@ref]}.}

\item{auth_token}{To install from a private repo, generate a personal
access token (PAT) in \url{https://github.com/settings/tokens} and
supply to this argument. This is safer than using a password because
you can easily delete a PAT without affecting any others. Defaults to
the \code{GITHUB_PAT} environment variable.}

\item{host}{GitLab API host to use. Override with your GitLab enterprise
hostname, for example, \code{"gitlab.hostname.com"}.}

\item{quiet}{if \code{TRUE} suppresses output from this function.}

\item{...}{Other arguments passed on to \code{\link{install}}.}
}
\description{
This function is vectorised on \code{repo} so you can install multiple
packages in a single command. Like other remotes the repository will skip
installation if `force == FALSE` (the default) and the remote state has
not changed since the previous installation.
}
\examples{
\dontrun{
install_github("jimhester/covr")
}
}
\seealso{
Other package installation: \code{\link{install_bioc}},
  \code{\link{install_bitbucket}},
  \code{\link{install_cran}}, \code{\link{install_github}},
  \code{\link{install_git}}, \code{\link{install_svn}},
  \code{\link{install_url}}, \code{\link{install_version}},
  \code{\link{install}}, \code{\link{uninstall}}
}
