% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/install-local.r
\name{install_local}
\alias{install_local}
\title{Install a package from a local file}
\usage{
install_local(path, subdir = NULL, ..., quiet = FALSE)
}
\arguments{
\item{path}{path to local directory, or compressed file (tar, zip, tar.gz
tar.bz2, tgz2 or tbz)}

\item{subdir}{subdirectory within url bundle that contains the R package.}

\item{...}{Other arguments passed on to \code{\link{install.packages}}.}

\item{quiet}{if \code{TRUE} suppresses output from this function.}
}
\description{
This function is vectorised so you can install multiple packages in
a single command.
}
\examples{
\dontrun{
dir <- tempfile()
dir.create(dir)
pkg <- download.packages("testthat", dir, type = "source")
install_local(pkg[, 2])

# Use `update_packages()` to upgrade the version of a package installed by
any of the `install_*()` functions.
update_packages("testthat")
}
}
\seealso{
\code{\link{update_packages}} to update installed packages from the
  source location
}
