% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/install-url.r
\name{install_url}
\alias{install_url}
\title{Install a package from a url}
\usage{
install_url(url, subdir = NULL, config = list(), ..., quiet = FALSE)
}
\arguments{
\item{url}{location of package on internet. The url should point to a
zip file, a tar file or a bzipped/gzipped tar file.}

\item{subdir}{subdirectory within url bundle that contains the R package.}

\item{config}{additional configuration argument (e.g. proxy,
authentication) passed on to \code{\link[httr]{GET}}.}

\item{...}{Other arguments passed on to \code{\link{install.packages}}.}

\item{quiet}{if \code{TRUE} suppresses output from this function.}
}
\description{
This function is vectorised so you can install multiple packages in
a single command.
}
\examples{
\dontrun{
install_url("https://github.com/hadley/stringr/archive/master.zip")

# Use `update_packages()` to upgrade the version of a package installed by
any of the `install_*()` functions.
update_packages("stringr")
}
}
\seealso{
\code{\link{update_packages}} to update installed packages from the
  source location

Other package installation: \code{\link{install_bioc}},
  \code{\link{install_bitbucket}},
  \code{\link{install_cran}}, \code{\link{install_github}},
  \code{\link{install_gitlab}}, \code{\link{install_git}},
  \code{\link{install_svn}}, \code{\link{install_version}},
  \code{\link{install}}, \code{\link{uninstall}}
}
\concept{package installation}
