\name{load_deps}
\alias{load_deps}
\title{Load dependencies.}
\usage{load_deps(pkg)}

\description{
  Load dependencies.
}

\details{
  Load all package dependencies as described in
  \code{DESCRIPTION}.
}
\keyword{programming}
\arguments{
  \item{pkg}{package description, can be path or package name.  See
\code{\link{as.package}} for more information}
}
