% Generated by roxygen2 (4.0.2): do not edit by hand
\name{load_dll}
\alias{load_dll}
\title{Load a compiled DLL}
\usage{
load_dll(pkg = ".")
}
\arguments{
\item{pkg}{package description, can be path or package name.  See
\code{\link{as.package}} for more information}
}
\description{
Load a compiled DLL
}
\keyword{programming}

