% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/make.R
\name{make}
\alias{make}
\alias{make_geteasytargets}
\alias{make_gettargets}
\title{Interface to `make` for package building}
\usage{
make(target = "all", path = "./", makefile = "Makefile", ...)

make_gettargets(makefile = "Makefile")

make_geteasytargets(makefile = "Makefile")
}
\arguments{
\item{target}{Makefile target, use \code{make_geteasytargets} to get a list of possible targets.}

\item{path}{path where makefile resides}

\item{makefile}{the Makefile to be used}

\item{...}{further parameters are handed to \code{\link[base]{system2}}.}
}
\description{
\code{make} calls make for the specified target.

\code{make_gettargets} extracts the targets from the makefile,
\code{make_geteasytargets} returns only the "easy" target names,
non-complex (no "%") targets that are also not filenames.
}
\examples{
\dontrun{
make ("build")
}
\dontrun{
make_gettargets ()
make_geteasytargets ()
}
}

