% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/missing-s3.R
\name{missing_s3}
\alias{missing_s3}
\title{Find missing s3 exports}
\usage{
missing_s3(pkg = ".")
}
\arguments{
\item{pkg}{The package to use, can be a file path to the package or a
package object.  See \code{\link[=as.package]{as.package()}} for more information.}
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#deprecated}{\figure{lifecycle-deprecated.svg}{options: alt='[Deprecated]'}}}{\strong{[Deprecated]}}

\code{missing_s3()} is deprecated because roxygen2 now provides the same
functionality. Run \code{devtools::document()} and look for
\code{"Missing documentation for S3 method"} warnings.
}
\keyword{internal}
