\name{on_path}
\alias{on_path}
\title{Test if an object is on the path}
\usage{
  on_path(...)
}
\arguments{
  \item{...}{character vectors indicating the exacutables
  to check for on the path.}
}
\value{
  logical TRUE of FALSE for each element of \code{...}
  indicating if it was found on the system path.
}
\description{
  Test if an object is on the path
}

