% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package.R
\name{package_file}
\alias{package_file}
\title{Find file in a package}
\usage{
package_file(..., path = ".")
}
\arguments{
\item{...}{Components of the path.}

\item{path}{Place to start search for package directory.}
}
\description{
It always starts by walking up the path until it finds the root directory,
i.e. a directory containing \code{DESCRIPTION}. If it cannot find the root
directory, or it can't find the specified path, it will throw an error.
}
\examples{
\dontrun{
package_file("figures", "figure_1")
}
}
\keyword{internal}
