\name{parse_deps}
\alias{parse_deps}
\title{Parse dependencies.}
\usage{
  parse_deps(string)
}
\value{
  list of two character vectors: \code{name} package names,
  and \code{version} package versions. If version is not
  specified, it will be stored as NA.
}
\description{
  Parse dependencies.
}
\keyword{internal}

