\name{pkg_imports_env}
\alias{pkg_imports_env}
\title{Package imports environment
Contains objects imported from other packages. Is a child of R_GlobalEnv
and is the parent of the package namespace environment.}
\usage{
  pkg_imports_env(pkg = NULL)
}
\description{
  Package imports environment Contains objects imported
  from other packages. Is a child of R_GlobalEnv and is the
  parent of the package namespace environment.
}

