\name{print.envlist}
\alias{print.envlist}
\title{Print an \code{envlist}}
\usage{
  print.envlist(x, name = TRUE, path = FALSE, ...)
}
\arguments{
  \item{name}{If \code{TRUE} (the default), print the
  \code{name} attribute of each environment.}

  \item{path}{If \code{TRUE}, print the \code{path}
  attribute of each environment.}
}
\description{
  Print an \code{envlist}
}

