% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/R.r
\name{RCMD}
\alias{RCMD}
\title{Run R CMD xxx from within R}
\usage{
RCMD(cmd, options, path = tempdir(), env_vars = NULL, ...)
}
\arguments{
\item{cmd}{one of the R tools available from the R CMD interface.}

\item{options}{a charater vector of options to pass to the command}

\item{path}{the directory to run the command in.}

\item{env_vars}{environment variables to set before running the command.}

\item{...}{additional arguments passed to \code{\link{system_check}}}
}
\value{
\code{TRUE} if the command succeeds, throws an error if the command
fails.
}
\description{
Run R CMD xxx from within R
}

