% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package-deps.R, R/pkgbuild.R, R/pkgload.R,
%   R/session-info.R
\docType{import}
\name{reexports}
\alias{reexports}
\alias{parse_deps}
\alias{check_dep_version}
\alias{with_debug}
\alias{clean_dll}
\alias{has_devel}
\alias{find_rtools}
\alias{is_loading}
\alias{unload}
\alias{session_info}
\alias{package_info}
\title{Objects exported from other packages}
\keyword{internal}
\description{
These objects are imported from other packages. Follow the links
below to see their documentation.

\describe{
  \item{pkgbuild}{\code{\link[pkgbuild:clean_dll]{clean_dll()}}, \code{\link[pkgbuild:find_rtools]{find_rtools()}}, \code{\link[pkgbuild:has_devel]{has_devel()}}, \code{\link[pkgbuild:with_debug]{with_debug()}}}

  \item{pkgload}{\code{\link[pkgload:check_dep_version]{check_dep_version()}}, \code{\link[pkgload:is_loading]{is_loading()}}, \code{\link[pkgload:parse_deps]{parse_deps()}}, \code{\link[pkgload:unload]{unload()}}}

  \item{sessioninfo}{\code{\link[sessioninfo:package_info]{package_info()}}, \code{\link[sessioninfo:session_info]{session_info()}}}
}}

