% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/check-devtools.r
\name{release_checks}
\alias{release_checks}
\title{Custom devtools release checks.}
\usage{
release_checks(pkg = ".", built_path = NULL)
}
\arguments{
\item{pkg}{package description, can be path or package name.  See
\code{\link{as.package}} for more information. If the \code{DESCRIPTION}
file does not exist, it is created using \code{\link{create_description}}.}
}
\description{
This function performs additional checks prior to release. It is called
automatically by \code{\link{release}()}.
}
\keyword{internal}

