\name{repositories}
\alias{repositories}
\title{Loads a package's specified repositories.}
\usage{
  repositories(pkg = NULL)
}
\arguments{
  \item{pkg}{package description, can be path or package
  name.  See \code{\link{as.package}} for more information}
}
\description{
  Loads a package's specified repositories.
}

