% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/revdep-email.R
\name{revdep_email}
\alias{revdep_email}
\title{Experimental email notification system.}
\usage{
revdep_email(pkg = ".", date, version, author = getOption("devtools.name"),
  draft = TRUE, unsent = NULL, template = "revdep/email.md",
  only_problems = TRUE)
}
\arguments{
\item{pkg}{Path to package. Defaults to current directory.}

\item{date}{Date package will be submitted to CRAN}

\item{version}{Version which will be used for the CRAN submission (usually
different from the current package version)}

\item{author}{Name used to sign email}

\item{draft}{If \code{TRUE}, creates as draft email; if \code{FALSE},
sends immediately.}

\item{unsent}{If some emails fail to send, in a previous}

\item{template}{Path of template to use}

\item{only_problems}{Only inform authors with problems?}
}
\description{
This currently assumes that you use github and gmail, and you have a
\code{revdep/email.md} email template.
}
\keyword{internal}
