% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/r-hub.R
\name{rhub_check}
\alias{rhub_check}
\title{Build and check package on r-hub}
\usage{
rhub_check(pkg = ".", platform = NULL, interactive = FALSE, ...)
}
\arguments{
\item{pkg}{package description, can be path or package name.  See
\code{\link{as.package}} for more information}

\item{platform}{the r-hub platform to run the check on. See
\code{\link[rhub]{platforms}} for the current platform list.
if \code{NULL}, then the desired platform can be selected
interactively.}

\item{interactive}{whether to show the status of the build
interactively. R-hub will send an email to the package maintainer's
email address, regardless of whether the check is interactive or not.}

\item{...}{extra arguments, passed to \code{\link{build}}.}
}
\value{
a handle that can be used to query the build status
  with \code{\link{rhub_check_status}}.
}
\description{
It runs \code{\link{build}} on the package, with the arguments specified
in \code{args}, and then submits it to the r-hub builder at
\url{https://builder.r-hub.io}. The \code{interactive} option controls
whether the function waits for the check output. Regardless, after the
check is complete, r-hub sends an email with the results to the package
maintainer.
}
\section{About email validation on r-hub}{

To build and check R packages on r-hub, you need to validate your
email address. This is because r-hub sends out emails about build
results. See more at \code{\link[rhub]{validate_email}}.
}

