% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/r-hub.R
\name{rhub_check_status}
\alias{rhub_check_status}
\title{Check status of an r-hub check submission}
\usage{
rhub_check_status(handle)
}
\arguments{
\item{handle}{the handle returned by \code{\link{rhub_check}},
or an r-hub job status URL, as printed by \code{\link{rhub_check}}.}
}
\value{
A list containing the status of the build, which can be
  \sQuote{preparing}, \sQuote{in progress}, \sQuote{success} and
  \sQuote{error}.
}
\description{
Check status of an r-hub check submission
}

