% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/save-all.R
\name{save_all}
\alias{save_all}
\title{Save all documents in an active IDE session}
\usage{
save_all()
}
\description{
Helper function wrapping IDE-specific calls to save all documents in the
active session. In this form, callers of \code{save_all()} don't need to
execute any IDE-specific code. This function can be extended to include
other IDE implementations of their equivalent
\code{rstudioapi::documentSaveAll()} methods.
}
