\name{save_package_description}
\alias{save_package_description}
\title{Writes the DESCRIPTION parameters to the given package.}
\usage{
  save_package_description(path, description)
}
\arguments{
  \item{path}{to the package}

  \item{description}{names vector of description values}
}
\description{
  Writes the DESCRIPTION parameters to the given package.
}

