% Generated by roxygen2 (4.0.1): do not edit by hand
\name{session_info}
\alias{session_info}
\title{Print session information}
\usage{
session_info()
}
\description{
This function prints out the same information as
\code{\link[utils]{sessionInfo}()}, with one difference: it also prints out
the GithubSHA1 value of loaded packages, if present.
}

