% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spell-check.R
\name{spell_check}
\alias{spell_check}
\title{Spell checking}
\usage{
spell_check(pkg = ".", ignore = character())
}
\arguments{
\item{pkg}{package description, can be path or package name.  See
\code{\link{as.package}} for more information}

\item{ignore}{character vector with words to ignore. See
\code{\link[hunspell:hunspell]{hunspell}} for more information}
}
\description{
Runs a spell check on text fields in the package description file and
manual pages.
}

