% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/release.R
\name{submit_cran}
\alias{submit_cran}
\title{Submit a package to CRAN}
\usage{
submit_cran(pkg = ".", args = NULL)
}
\arguments{
\item{pkg}{The package to use, can be a file path to the package or a
package object.  See \code{\link[=as.package]{as.package()}} for more information.}

\item{args}{An optional character vector of additional command
line arguments to be passed to \verb{R CMD build}.}
}
\description{
This submits your package to CRAN using the web-form submission process.
To complete the submission you will need respond to the email sent to the
maintainer email address.

We generally recommend using this part of the process defined by
\code{\link[usethis:use_release_issue]{use_release_issue()}}; this process maximizes the chances of a successful
submission.
}
\keyword{internal}
