% Generated by roxygen2 (4.0.1): do not edit by hand
\name{svn_path}
\alias{svn_path}
\title{Retrieve the current running path of the svn binary.}
\usage{
svn_path(svn_binary_name = NULL)
}
\arguments{
\item{svn_binary_name}{The name of the binary depending on the OS.}
}
\description{
Retrieve the current running path of the svn binary.
}

