% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/system.R
\name{system_check}
\alias{system_check}
\title{Run a system command and check if it succeeds.}
\usage{
system_check(cmd, args = character(), env_vars = character(),
  path = ".", quiet = FALSE, throw = TRUE, ...)
}
\arguments{
\item{cmd}{Command to run. Will be quoted by \code{\link[=shQuote]{shQuote()}}.}

\item{args}{A character vector of arguments.}

\item{env_vars}{A named character vector of environment variables.}

\item{path}{Path in which to execute the command}

\item{quiet}{If \code{FALSE}, the command to be run will be echoed.}

\item{throw}{If \code{TRUE}, will throw an error if the command fails
(i.e. the return value is not 0).}

\item{...}{additional arguments passed to \code{\link[base:system]{base::system()}}}
}
\value{
The exit status of the command, invisibly.
}
\description{
Run a system command and check if it succeeds.
}
\keyword{internal}
