% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uninstall.R
\name{uninstall}
\alias{uninstall}
\title{Uninstall a local development package}
\usage{
uninstall(pkg = ".", unload = TRUE, quiet = FALSE, lib = .libPaths()[[1]])
}
\arguments{
\item{pkg}{The package to use, can be a file path to the package or a
package object.  See \code{\link[=as.package]{as.package()}} for more information.}

\item{unload}{if \code{TRUE} (the default), ensures the package is unloaded, prior
to uninstalling.}

\item{quiet}{If \code{TRUE}, suppress output.}

\item{lib}{a character vector giving the library directories to remove the
    packages from.  If missing, defaults to the first element in
    \code{\link{.libPaths}()}.}
}
\description{
Uses \code{remove.packages()} to uninstall the package. To uninstall a package
from a non-default library, use in combination with \code{\link[withr:with_libpaths]{withr::with_libpaths()}}.
}
\seealso{
\code{\link[=with_debug]{with_debug()}} to install packages with debugging flags set.

Other package installation:
\code{\link[=install]{install()}}
}
\concept{package installation}
