% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update.R
\name{update_packages}
\alias{update_packages}
\title{Update packages that are missing or out-of-date.}
\usage{
update_packages(pkgs = NULL, repos = getOption("repos"),
  type = getOption("pkgType"), include_remote = FALSE, ...)
}
\arguments{
\item{pkgs}{Character vector of packages to update. Leave out to update all
installed packages.}

\item{repos}{character vector, the base URL(s) of the repositories to use.}

\item{type}{character, indicating the type of package to download and install.}

\item{include_remote}{A logical flag to update packages installed from git
and GitHub repositories.}

\item{...}{Additional parameters to pass to \code{update_remotes}
(e.g. \code{ask}).}
}
\description{
Works similarly to \code{install.packages()} but doesn't install packages
that are already installed, and also upgrades out dated dependencies. Can
also update packages installed from remote repositories. Currently only
works with packages installed from git and GitHub repositories. This feature
does not require git to be installed.
}
\examples{
\dontrun{
update_packages("ggplot2")

# install_git("git://github.com/hadley/stringr.git")
update_remotes("stringr")

# A mix of CRAN and git-installed packages:
update_packages(c("plyr", "ggplot2", "stringr"), ask = FALSE)
}
}
\seealso{
\code{\link{package_deps}} to see which packages are out of date/
  missing.
}

