% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update.R
\name{update_remotes}
\alias{update_remotes}
\title{Update packages installed from remote repositories.}
\usage{
update_remotes(pkgs = NULL, ask = TRUE)
}
\arguments{
\item{pkgs}{Character vector of package names. If omitted,
will find all packages installed from remote repositories.}

\item{ask}{Forces the process to ask the user for confirmation.}
}
\description{
It is vectorised so you can update multiple packages
  with a single command. Currently only works with packages
  installed from git and GitHub. You do not need to have git installed.
}
\examples{
\dontrun{
# install_git("git://github.com/hadley/stringr.git")
update_remotes("stringr")

update_remotes(ask = FALSE)
}
}

