% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/infrastructure-git.R
\name{use_git}
\alias{use_git}
\title{Initialise a git repository.}
\usage{
use_git(message = "Initial commit", pkg = ".")
}
\arguments{
\item{message}{Message to use for first commit.}

\item{pkg}{Path to package. See \code{\link{as.package}} for more
information.}
}
\description{
Initialise a git repository.
}
\examples{
\dontrun{use_git()}
}
\seealso{
Other git infrastructure: \code{\link{use_git_hook}};
  \code{\link{use_github}}
}

