% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/infrastructure-git.R
\name{use_git_hook}
\alias{use_git_hook}
\title{Add a git hook.}
\usage{
use_git_hook(hook, script, pkg = ".")
}
\arguments{
\item{hook}{Hook name. One of "pre-commit", "prepare-commit-msg",
"commit-msg", "post-commit", "applypatch-msg", "pre-applypatch",
"post-applypatch", "pre-rebase", "post-rewrite", "post-checkout",
"post-merge", "pre-push", "pre-auto-gc".}

\item{script}{Text of script to run}

\item{pkg}{Path to package. See \code{\link{as.package}} for more
information.}
}
\description{
Add a git hook.
}
\seealso{
Other git infrastructure: \code{\link{use_github}};
  \code{\link{use_git}}
}
\keyword{internal}

