% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/infrastructure-git.R
\name{use_github_links}
\alias{use_github_links}
\title{Add GitHub links to DESCRIPTION.}
\usage{
use_github_links(pkg = ".", host = "api.github.com")
}
\arguments{
\item{pkg}{Path to package. See \code{\link{as.package}} for more
information.}

\item{host}{GitHub API host to use. Override with your GitHub enterprise
hostname, for example, "github.hostname.com/api/v3".}
}
\description{
Populates the URL and BugReports fields of DESCRIPTION with
\code{https://github.com/<USERNAME>/<REPO>} AND
\code{https://github.com/<USERNAME>/<REPO>/issues}, respectively, unless
those fields already exist.
}
\seealso{
Other git infrastructure: \code{\link{use_git_hook}},
  \code{\link{use_github}}, \code{\link{use_git}}
}
\keyword{internal}

