% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/infrastructure.R
\name{use_news_md}
\alias{use_news_md}
\title{Use NEWS.md}
\usage{
use_news_md(pkg = ".")
}
\arguments{
\item{pkg}{package description, can be path or package name.  See
\code{\link{as.package}} for more information}
}
\description{
This creates \code{NEWS.md} from a template.
}
\seealso{
Other infrastructure: \code{\link{infrastructure}},
  \code{\link{use_build_ignore}},
  \code{\link{use_data_raw}}, \code{\link{use_data}},
  \code{\link{use_package}}, \code{\link{use_readme_rmd}}
}
