% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/infrastructure.R
\name{use_package}
\alias{use_package}
\title{Use specified package.}
\usage{
use_package(package, type = "Imports", pkg = ".", version = NULL,
  compare = c(">=", ">", "==", "<=", "<"))
}
\arguments{
\item{package}{Name of package to depend on.}

\item{type}{Type of dependency: must be one of "Imports", "Suggests",
"Depends", "Suggests", "Enhances", or "LinkingTo" (or unique abbreviation)}

\item{pkg}{package description, can be path or package name. See
\code{\link{as.package}} for more information.}

\item{version}{Value indicating whether \code{package}'s version should be included
in \code{pkg}'s DESCRIPTION. Either \code{NULL}, \code{TRUE}, or a valid version
string. If \code{version} is \code{TRUE}, \code{package}'s current version
will be used. See examples.}

\item{compare}{The comparator used for \code{package}'s version. All valid
CRAN comparators (i.e. \code{==}, \code{>=}, \code{<=}, \code{>}, and
\code{<}) are accepted, but not considered unless \code{version} is
non-\code{NULL}.}
}
\description{
This adds a dependency to DESCRIPTION and offers a little advice
about how to best use it.
}
\examples{
\dontrun{
use_package("ggplot2")
use_package("dplyr", "suggests")
use_package("devtools", version = TRUE) # "devtools (>= 1.10.0.9000)"
use_package("devtools", version = TRUE, compare = "==") # "devtools (== 1.10.0.9000)"
use_package("devtools", version = "1.10") # "devtools (>= 1.10)"
use_package("devtools", version = "1.10", compare = ">") # "devtools (> 1.10)"
}
}
\seealso{
Other infrastructure: \code{\link{infrastructure}},
  \code{\link{use_build_ignore}},
  \code{\link{use_data_raw}}, \code{\link{use_data}},
  \code{\link{use_news_md}}, \code{\link{use_readme_rmd}}
}

