% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/infrastructure.R
\name{use_readme_rmd}
\alias{use_readme_rmd}
\alias{use_readme_md}
\title{Create README files.}
\usage{
use_readme_rmd(pkg = ".")

use_readme_md(pkg = ".")
}
\arguments{
\item{pkg}{package description, can be path or package name.  See
\code{\link{as.package}} for more information}
}
\description{
Creates skeleton README files with sections for
\itemize{
\item a high-level description of the package and its goals
\item R code to install from GitHub, if GitHub usage detected
\item a basic example
}
Use \code{Rmd} if you want a rich intermingling of code and data. Use
\code{md} for a basic README. \code{README.Rmd} will be automatically
added to \code{.Rbuildignore}. The resulting README is populated with default
YAML frontmatter and R fenced code blocks (\code{md}) or chunks (\code{Rmd}).
}
\examples{
\dontrun{
use_readme_rmd()
use_readme_md()
}
}
\seealso{
Other infrastructure: \code{\link{infrastructure}},
  \code{\link{use_build_ignore}},
  \code{\link{use_data_raw}}, \code{\link{use_data}},
  \code{\link{use_news_md}}, \code{\link{use_package}}
}
