% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wd.R
\name{wd}
\alias{wd}
\title{Set working directory}
\usage{
wd(pkg = ".", path = "")
}
\arguments{
\item{pkg}{The package to use, can be a file path to the package or a
package object.  See \code{\link[=as.package]{as.package()}} for more information.}

\item{path}{path within package. Leave empty to change working directory
to package directory.}
}
\description{
Set working directory
}
