\name{write_description}
\alias{write_description}
\title{Writes the DESCRIPTION parameters to the given package.}
\usage{
  write_description(path, description)
}
\arguments{
  \item{path}{to the package}

  \item{description}{names vector of description values}
}
\description{
  Writes the DESCRIPTION parameters to the given package.
}

