% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package.R
\name{as.package}
\alias{as.package}
\title{Coerce input to a package}
\usage{
as.package(x = NULL, create = deprecated())
}
\arguments{
\item{x}{object to coerce to a package}

\item{create}{\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#deprecated}{\figure{lifecycle-deprecated.svg}{options: alt='[Deprecated]'}}}{\strong{[Deprecated]}} Hasn't worked for some time.}
}
\description{
Possible specifications of package:
\itemize{
\item path
\item package object
}
}
\keyword{internal}
