\name{as.package}
\alias{as.package}
\title{Coerce input to a package.}
\usage{as.package(x)}

\description{
  Coerce input to a package.
}

\details{
  Possibile specifications of package:

  \itemize{ \item path \item name (lookup in .Rpackages)
  \item package object }
}
\arguments{
  \item{x}{object to coerce to a package}
}
