% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bash.R
\name{bash}
\alias{bash}
\title{Open bash shell in package directory}
\usage{
bash(pkg = ".")
}
\arguments{
\item{pkg}{The package to use, can be a file path to the package or a
package object.  See \code{\link[=as.package]{as.package()}} for more information.}
}
\description{
Open bash shell in package directory
}
