% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bash.R
\name{bash}
\alias{bash}
\title{Open bash shell in package directory}
\usage{
bash(pkg = ".")
}
\arguments{
\item{pkg}{The package to use, can be a file path to the package or a
package object.  See \code{\link[=as.package]{as.package()}} for more information.}
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#deprecated}{\figure{lifecycle-deprecated.svg}{options: alt='[Deprecated]'}}}{\strong{[Deprecated]}}

\code{bash()} is deprecated because we no longer use or recommend this workflow.
Open bash shell in package directory
}
\keyword{internal}
