% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/bash.r
\name{bash}
\alias{bash}
\title{Open bash shell in package directory.}
\usage{
bash(pkg = ".")
}
\arguments{
\item{pkg}{package description, can be path or package name.  See
\code{\link{as.package}} for more information}
}
\description{
Open bash shell in package directory.
}

