% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build-readme.R
\name{build_readme}
\alias{build_readme}
\title{Build a Rmarkdown README for a package}
\usage{
build_readme(path = ".", quiet = TRUE, ...)
}
\arguments{
\item{path}{path to the package to build the readme.}

\item{quiet}{If \code{TRUE}, suppress output.}

\item{...}{additional arguments passed to \code{\link[rmarkdown:render]{rmarkdown::render()}}}
}
\description{
\code{build_readme()} is a wrapper around \code{\link[rmarkdown:render]{rmarkdown::render()}}, it generates the
README.md from a README.Rmd file.
}
