% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build-site.R
\name{build_site}
\alias{build_site}
\title{Execute \pkg{pkgdown} build_site in a package}
\usage{
build_site(path = ".", ...)
}
\arguments{
\item{path}{path to the package to build the static HTML.}

\item{...}{additional arguments passed to [pkgdown::build_site()]}
}
\description{
`build_site()` is a shortcut for [pkgdown::build_site()], it generates the
  static HTML documentation.
}
