% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build-site.R
\name{build_site}
\alias{build_site}
\title{Run \code{pkgdown::build_site()}}
\usage{
build_site(path = ".", ...)
}
\arguments{
\item{path}{Path to the package to build the static HTML.}

\item{...}{Additional arguments passed to \code{\link[pkgdown:build_site]{pkgdown::build_site()}}.}
}
\description{
This is a thin wrapper around \code{\link[pkgdown:build_site]{pkgdown::build_site()}}, used for generating
static HTML documentation. Learn more at \url{https://pkgdown.r-lib.org}.
}
