% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vignettes.r
\name{build_vignettes}
\alias{build_vignettes}
\title{Build package vignettes.}
\usage{
build_vignettes(pkg = ".", dependencies = "VignetteBuilder", clean = TRUE,
  upgrade = FALSE, quiet = TRUE)
}
\arguments{
\item{pkg}{package description, can be path or package name.  See
\code{\link{as.package}} for more information}

\item{dependencies}{\code{logical} indicating to also install uninstalled
packages which this \code{pkg} depends on/links to/suggests. See
argument \code{dependencies} of \code{\link{install.packages}}.}

\item{clean}{Remove all files generated by the build, even if there
    were copies there before.}

\item{upgrade}{If \code{TRUE}, also upgrade any of out date dependencies.}

\item{quiet}{If \code{TRUE}, suppresses most output. Set to \code{FALSE}
if you need to debug.}
}
\description{
Builds package vignettes using the same algorithm that \code{R CMD build}
does. This means including non-Sweave vignettes, using makefiles (if
present), and copying over extra files. You need to ensure that these
files are not included in the built package - ideally they should not
be checked into source, or at least excluded with \code{.Rbuildignore}
}
\seealso{
\code{\link{clean_vignettes}} to remove the pdfs in
  \file{inst/doc} created from vignettes

\code{\link{clean_vignettes}} to remove build tex/pdf files.
}
\keyword{programming}
