% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vignettes.r
\name{build_vignettes}
\alias{build_vignettes}
\title{Build package vignettes.}
\usage{
build_vignettes(pkg = ".", dependencies = "VignetteBuilder",
  clean = TRUE, upgrade = TRUE, quiet = TRUE, install = TRUE,
  keep_md = TRUE)
}
\arguments{
\item{pkg}{package description, can be path or package name.  See
\code{\link{as.package}} for more information}

\item{dependencies}{Which dependencies do you want to check?
  Can be a character vector (selecting from "Depends", "Imports",
   "LinkingTo", "Suggests", or "Enhances"), or a logical vector.

  \code{TRUE} is shorthand for "Depends", "Imports", "LinkingTo" and
  "Suggests". \code{NA} is shorthand for "Depends", "Imports" and "LinkingTo"
  and is the default. \code{FALSE} is shorthand for no dependencies (i.e.
  just check this package, not its dependencies).}

\item{clean}{Remove all files generated by the build, even if there
    were copies there before.}

\item{upgrade}{If \code{TRUE}, also upgrade any of out date dependencies.}

\item{quiet}{If \code{TRUE}, suppresses most output. Set to \code{FALSE}
if you need to debug.}

\item{install}{If \code{TRUE}, install the package before building
vignettes.}

\item{keep_md}{If \code{TRUE}, move md intermediates as well as rendered
outputs. Most useful when using the `keep_md` YAML option for Rmarkdown
outputs. See
\url{https://bookdown.org/yihui/rmarkdown/html-document.html#keeping-markdown}.}
}
\description{
Builds package vignettes using the same algorithm that \code{R CMD build}
does. This means including non-Sweave vignettes, using makefiles (if
present), and copying over extra files. The files are copied in the 'doc'
directory and an vignette index is created in 'Meta/vignette.rds', as they
would be in a built package. 'doc' and 'Meta' are added to
\code{.Rbuildignore}, so will not be included in the built package. These
files can be checked into version control, so they can be viewed with
\code{browseVignettes()} and \code{vignette()} if the package has been
loaded with \code{load_all()} without needing to re-build them locally.
}
\seealso{
\code{\link{clean_vignettes}} to remove the pdfs in
  \file{doc} created from vignettes

\code{\link{clean_vignettes}} to remove build tex/pdf files.
}
\keyword{programming}
