% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vignettes.R
\name{build_vignettes}
\alias{build_vignettes}
\title{Build package vignettes}
\usage{
build_vignettes(
  pkg = ".",
  dependencies = "VignetteBuilder",
  clean = TRUE,
  upgrade = "never",
  quiet = FALSE,
  install = TRUE,
  keep_md = TRUE
)
}
\arguments{
\item{pkg}{The package to use, can be a file path to the package or a
package object.  See \code{\link[=as.package]{as.package()}} for more information.}

\item{clean}{Remove all files generated by the build, even if there
    were copies there before.}

\item{quiet}{If \code{TRUE}, suppresses most output. Set to \code{FALSE}
if you need to debug.}

\item{install}{If \code{TRUE}, install the package before building
vignettes.}

\item{keep_md}{If \code{TRUE}, move md intermediates as well as rendered
outputs. Most useful when using the \code{keep_md} YAML option for Rmarkdown
outputs. See
\url{https://bookdown.org/yihui/rmarkdown/html-document.html#keeping-markdown}.}
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#deprecated}{\figure{lifecycle-deprecated.svg}{options: alt='[Deprecated]'}}}{\strong{[Deprecated]}}

\code{build_vignettes()} is deprecated because we no longer recommend that you
build vignettes in this way, because it leaves build artifacts in your
development directory. Instead, use \code{\link[pkgdown:build_articles]{pkgdown::build_article()}} to
render articles locally for preview and polishing.
}
\keyword{internal}
