% Generated by roxygen2 (4.0.2): do not edit by hand
\name{build_win}
\alias{build_win}
\title{Build windows binary package.}
\usage{
build_win(pkg = ".", version = c("R-release", "R-devel"), args = NULL,
  quiet = FALSE)
}
\arguments{
\item{pkg}{package description, can be path or package name.  See
\code{\link{as.package}} for more information}

\item{version}{directory to upload to on the win-builder, controlling
which version of R is used to build the package. Possible options are
listed on \url{http://win-builder.r-project.org/}. Defaults to the
released version of R.}

\item{args}{An optional character vector of additional command
line arguments to be passed to \code{R CMD build} if \code{binary = FALSE},
or \code{R CMD install} if \code{binary = TRUE}.}

\item{quiet}{if \code{TRUE} suppresses output from this function.}
}
\description{
This function works by bundling source package, and then uploading to
\url{http://win-builder.r-project.org/}.  Once building is complete you'll
receive a link to the built package in the email address listed in the
maintainer field.  It usually takes around 30 minutes. As a side effect,
win-build also runs \code{R CMD check} on the package, so \code{build_win}
is also useful to check that your package is ok on windows.
}
\seealso{
Other build functions: \code{\link{build}}
}

