\name{check}
\alias{check}
\title{Build and check a package, cleaning up automatically on success.}
\usage{check(pkg)}

\description{
  Build and check a package, cleaning up automatically on
  success.
}
\arguments{
  \item{pkg}{package description, can be path or package name.  See
\code{\link{as.package}} for more information}
}
