% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check-mac.R
\name{check_mac_release}
\alias{check_mac_release}
\alias{check_mac_devel}
\title{Check a package on macOS}
\usage{
check_mac_release(
  pkg = ".",
  dep_pkgs = character(),
  args = NULL,
  manual = TRUE,
  quiet = FALSE,
  ...
)

check_mac_devel(
  pkg = ".",
  dep_pkgs = character(),
  args = NULL,
  manual = TRUE,
  quiet = FALSE,
  ...
)
}
\arguments{
\item{pkg}{The package to use, can be a file path to the package or a
package object.  See \code{\link[=as.package]{as.package()}} for more information.}

\item{dep_pkgs}{Additional custom dependencies to install prior to checking
the package.}

\item{args}{An optional character vector of additional command
line arguments to be passed to \verb{R CMD build} if \code{binary = FALSE},
or \verb{R CMD install} if \code{binary = TRUE}.}

\item{manual}{Should the manual be built?}

\item{quiet}{If \code{TRUE}, suppresses output.}

\item{...}{Additional arguments passed to \code{\link[pkgbuild:build]{pkgbuild::build()}}.}
}
\value{
The url with the check results (invisibly)
}
\description{
Check on either the released or development versions of R, using
\url{https://mac.r-project.org/macbuilder/}.
}
\seealso{
Other build functions: 
\code{\link{check_win}()}
}
\concept{build functions}
