% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check-mac.R
\name{check_mac_release}
\alias{check_mac_release}
\title{Check macOS package}
\usage{
check_mac_release(
  pkg = ".",
  dep_pkgs = character(),
  args = NULL,
  manual = TRUE,
  quiet = FALSE,
  ...
)
}
\arguments{
\item{pkg}{The package to use, can be a file path to the package or a
package object.  See \code{\link[=as.package]{as.package()}} for more information.}

\item{dep_pkgs}{Additional custom dependencies to install prior to checking the package.}

\item{args}{An optional character vector of additional command
line arguments to be passed to \verb{R CMD build} if \code{binary = FALSE},
or \verb{R CMD install} if \code{binary = TRUE}.}

\item{quiet}{If \code{TRUE}, suppresses output.}

\item{...}{Additional arguments passed to \code{\link[pkgbuild:build]{pkgbuild::build()}}.}
}
\value{
The url with the check results (invisibly)
}
\description{
This function works by bundling source package, and then uploading to
\url{https://mac.r-project.org/macbuilder/submit.html}.  This function returns a
link to the page with the check results.
}
\seealso{
Other build functions: 
\code{\link{check_rhub}()},
\code{\link{check_win}()}
}
\concept{build functions}
