% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/r-hub.R
\name{check_rhub}
\alias{check_rhub}
\title{Run CRAN checks for package on R-hub}
\usage{
check_rhub(
  pkg = ".",
  platforms = NULL,
  email = NULL,
  interactive = TRUE,
  build_args = NULL,
  ...
)
}
\arguments{
\item{pkg}{The package to use, can be a file path to the package or a
package object.  See \code{\link[=as.package]{as.package()}} for more information.}

\item{platforms}{R-hub platforms to run the check on.}

\item{email}{email address to notify.}

\item{interactive}{whether to show the status of the build.}

\item{build_args}{Arguments passed to \verb{R CMD build}.}

\item{...}{extra arguments, passed to \code{rhub::check_for_cran()}.}
}
\value{
a \code{rhub_check} object.
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#deprecated}{\figure{lifecycle-deprecated.svg}{options: alt='[Deprecated]'}}}{\strong{[Deprecated]}}

This function is deprecated since the underlying function
\code{rhub::check_for_cran()} is now deprecated and defunct. See \code{\link[rhub:rhubv2]{rhub::rhubv2}}
learn about the new check system, R-hub v2.
}
\keyword{internal}
