% Generated by roxygen2 (4.0.0): do not edit by hand
\name{clean_source}
\alias{clean_source}
\title{Sources an R file in a clean environment.}
\usage{
clean_source(path, quiet = FALSE)
}
\arguments{
\item{path}{path to R script}

\item{quiet}{If \code{FALSE}, the default, all input and output will be
displayed, as if you'd copied and paste the code.  If \code{TRUE}
only the final result and the any explicitly printed output will be
displayed.}
}
\description{
Opens up a fresh R environment and sources file, ensuring that it works
independently of the current working environment.
}

