\name{clean_vignettes}
\alias{clean_vignettes}
\title{Clean built vignettes.}
\usage{
  clean_vignettes(pkg = ".", tex = TRUE)
}
\arguments{
  \item{pkg}{package description, can be path or package
  name.  See \code{\link{as.package}} for more information}

  \item{tex}{if \code{TRUE} also removes all tex build
  artefacts}
}
\description{
  Clean built vignettes.
}

