% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vignettes.R
\name{clean_vignettes}
\alias{clean_vignettes}
\title{Clean built vignettes}
\usage{
clean_vignettes(pkg = ".")
}
\arguments{
\item{pkg}{The package to use, can be a file path to the package or a
package object.  See \code{\link[=as.package]{as.package()}} for more information.}
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#deprecated}{\figure{lifecycle-deprecated.svg}{options: alt='[Deprecated]'}}}{\strong{[Deprecated]}}

\code{clean_vignettes()} is deprecated because \code{\link[=build_vignettes]{build_vignettes()}} is deprecated.
}
\keyword{internal}
