% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create.R
\name{create}
\alias{create}
\title{Create a package}
\usage{
create(path, ...)
}
\arguments{
\item{path}{A path. If it exists, it is used. If it does not exist, it is
created, provided that the parent path exists.}

\item{...}{Additional arguments passed to \code{\link[usethis:create_package]{usethis::create_package()}}}
}
\value{
The path to the created package, invisibly.
}
\description{
Create a package
}
