% Generated by roxygen2 (4.0.0): do not edit by hand
\name{create_description}
\alias{create_description}
\title{Create a default DESCRIPTION file for a package.}
\usage{
create_description(path, extra = getOption("devtools.desc"), quiet = FALSE)
}
\arguments{
\item{path}{path to package root directory}

\item{extra}{a named list of extra options to add to \file{DESCRIPTION}.
Arguments that take a list}

\item{quiet}{if \code{TRUE}, suppresses output from this function.}
}
\description{
Create a default DESCRIPTION file for a package.
}
\details{
To set the default author and licenses, set \code{options}
\code{devtools.desc.author} and \code{devtools.desc.license}.  I use
\code{options(devtools.desc.author = '"Hadley Wickham <h.wickham@gmail.com> [aut,cre]"',
  devtools.desc.license = "GPL-3")}.
}

