% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/revdep-email.R, R/revdep-summarise.R,
%   R/revdep.R, R/zzz.r
\name{revdep_email}
\alias{revdep_email}
\alias{revdep_check_save_summary}
\alias{revdep_check_print_problems}
\alias{revdep_check}
\alias{revdep_check_resume}
\alias{revdep_check_reset}
\alias{devtools-deprecated}
\title{Deprecated Functions}
\usage{
revdep_email(pkg = ".", date, version, author = getOption("devtools.name"),
  draft = TRUE, unsent = NULL, template = "revdep/email.md",
  only_problems = TRUE)

revdep_check_save_summary(pkg = ".")

revdep_check_print_problems(pkg = ".")

revdep_check(pkg = ".", recursive = FALSE, ignore = NULL,
  dependencies = c("Depends", "Imports", "Suggests", "LinkingTo"),
  skip = character(), libpath = getOption("devtools.revdep.libpath"),
  srcpath = libpath, bioconductor = FALSE, type = getOption("pkgType"),
  threads = getOption("Ncpus", 1), env_vars = NULL, check_dir = NULL,
  install_dir = NULL, quiet_check = TRUE)

revdep_check_resume(pkg = ".", ...)

revdep_check_reset(pkg = ".")
}
\description{
These functions are Deprecated in this release of devtools, they will be
marked as Defunct and removed in a future version.
}
\details{
Instead of the revdep functions in devtools a better alternative is
\sQuote{revdepcheck::revdep_check()}.
}
\keyword{internal}
