% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/infrastructure.R, R/zzz.r
\name{use_coveralls}
\alias{devtools-deprecated}
\alias{use_coveralls}
\title{Deprecated Functions}
\usage{
use_coveralls(pkg = ".")
}
\description{
These functions are Deprecated in this release of devtools, they will be
marked as Defunct and removed in a future version.
}
\section{\code{use_coveralls}}{

Add coveralls to basic travis template to a package.
}
\keyword{internal}

