% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sitrep.R, R/zzz.R
\name{dr_devtools}
\alias{dr_devtools}
\alias{dr_github}
\alias{devtools-deprecated}
\title{Deprecated Functions}
\usage{
dr_devtools()

dr_github(path = ".")
}
\description{
These functions are Deprecated in this release of devtools, they will be
marked as Defunct and removed in a future version.
}
\keyword{internal}
