% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/with.r, R/zzz.r
\name{in_dir}
\alias{devtools-deprecated}
\alias{in_dir}
\alias{with_collate}
\alias{with_envvar}
\alias{with_lib}
\alias{with_libpaths}
\alias{with_locale}
\alias{with_makevars}
\alias{with_options}
\alias{with_par}
\alias{with_path}
\title{Deprecated Functions}
\usage{
in_dir(new, code)

with_collate(new, code)

with_envvar(new, code, action = "replace")

with_lib(new, code)

with_libpaths(new, code)

with_locale(new, code)

with_makevars(new, code, path = file.path("~", ".R", "Makevars"))

with_options(new, code)

with_par(new, code)

with_path(new, code, add = TRUE)
}
\description{
These functions are Deprecated in this release of devtools, they will be
marked as Defunct and removed in a future version.
}
\section{\code{in_dir}}{

working directory
}

\section{\code{with_collate}}{

collation order
}

\section{\code{with_envvar}}{

environmental variables
}

\section{\code{with_lib}}{

library paths, prepending to current libpaths
}

\section{\code{with_libpaths}}{

library paths, replacing current libpaths
}

\section{\code{with_locale}}{

any locale setting
}

\section{\code{with_makevars}}{

Temporarily change contents of an existing Makevars file.
}

\section{\code{with_options}}{

options
}

\section{\code{with_par}}{

graphics parameters
}

\section{\code{with_path}}{

PATH environment variable
}
\keyword{internal}

