% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check-cran.R, R/check-win.R, R/zzz.R
\name{check_cran}
\alias{check_cran}
\alias{build_win}
\alias{devtools-deprecated}
\title{Deprecated Functions}
\usage{
check_cran(pkgs, libpath = file.path(tempdir(), "R-lib"),
  srcpath = libpath, check_libpath = libpath, bioconductor = FALSE,
  type = getOption("pkgType"), threads = getOption("Ncpus", 1),
  check_dir = tempfile("check_cran"),
  install_dir = tempfile("check_cran_install"), env_vars = NULL,
  quiet_check = TRUE)

build_win(pkg = ".", version = c("R-devel", "R-release"))
}
\description{
These functions are Deprecated in this release of devtools, they will be
marked as Defunct and removed in a future version.
}
\keyword{internal}
