% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check-win.R, R/zzz.R
\name{build_win}
\alias{build_win}
\alias{devtools-deprecated}
\title{Deprecated Functions}
\usage{
build_win(pkg = ".", version = c("R-devel", "R-release"))
}
\description{
These functions are Deprecated in this release of devtools, they will be
marked as Defunct and removed in a future version.
}
\keyword{internal}
