% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\name{devtools-deprecated}
\alias{devtools-deprecated}
\title{Deprecated Functions}
\description{
These functions are Deprecated in this release of devtools, they will be
marked as Defunct and removed in a future version.
}
\keyword{internal}
