% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check-cran.r, R/create.r,
%   R/infrastructure-git.R, R/infrastructure.R, R/revdep-email.R,
%   R/revdep-summarise.R, R/revdep.R, R/zzz.r
\name{check_cran}
\alias{check_cran}
\alias{create}
\alias{setup}
\alias{create_description}
\alias{use_git}
\alias{use_github}
\alias{use_git_hook}
\alias{use_github_links}
\alias{use_testthat}
\alias{use_test}
\alias{use_rstudio}
\alias{use_vignette}
\alias{use_rcpp}
\alias{use_travis}
\alias{use_coverage}
\alias{use_appveyor}
\alias{use_package_doc}
\alias{use_package}
\alias{use_data}
\alias{use_data_raw}
\alias{use_build_ignore}
\alias{use_readme_rmd}
\alias{use_readme_md}
\alias{use_news_md}
\alias{use_revdep}
\alias{use_cran_comments}
\alias{use_code_of_conduct}
\alias{use_cran_badge}
\alias{use_mit_license}
\alias{use_gpl3_license}
\alias{use_dev_version}
\alias{revdep_email}
\alias{revdep_check_save_summary}
\alias{revdep_check_print_problems}
\alias{revdep_check}
\alias{revdep_check_resume}
\alias{revdep_check_reset}
\alias{devtools-deprecated}
\title{Deprecated Functions}
\usage{
check_cran(pkgs, libpath = file.path(tempdir(), "R-lib"), srcpath = libpath,
  check_libpath = libpath, bioconductor = FALSE,
  type = getOption("pkgType"), threads = getOption("Ncpus", 1),
  check_dir = tempfile("check_cran"),
  install_dir = tempfile("check_cran_install"), env_vars = NULL,
  quiet_check = TRUE)

create(path, description = getOption("devtools.desc"), check = FALSE,
  rstudio = TRUE, quiet = FALSE)

setup(path = ".", description = getOption("devtools.desc"), check = FALSE,
  rstudio = TRUE, quiet = FALSE)

create_description(path = ".", extra = getOption("devtools.desc"),
  quiet = FALSE)

use_git(message = "Initial commit", pkg = ".")

use_github(auth_token = github_pat(), private = FALSE, pkg = ".",
  host = "https://api.github.com", protocol = c("ssh", "https"),
  credentials = NULL, ...)

use_git_hook(hook, script, pkg = ".")

use_github_links(pkg = ".", auth_token = github_pat(),
  host = "https://api.github.com")

use_testthat(pkg = ".")

use_test(name, pkg = ".")

use_rstudio(pkg = ".")

use_vignette(name, pkg = ".")

use_rcpp(pkg = ".")

use_travis(pkg = ".", browse = interactive())

use_coverage(pkg = ".", type = c("codecov", "coveralls"))

use_appveyor(pkg = ".")

use_package_doc(pkg = ".")

use_package(package, type = "Imports", pkg = ".")

use_data(..., internal = FALSE, overwrite = FALSE, compress = "bzip2")

use_data_raw(pkg = ".")

use_build_ignore(files, escape = TRUE, pkg = ".")

use_readme_rmd(pkg = ".")

use_readme_md(pkg = ".")

use_news_md(pkg = ".")

use_revdep(pkg = ".")

use_cran_comments(pkg = ".")

use_code_of_conduct(pkg = ".")

use_cran_badge(pkg = ".")

use_mit_license(pkg = ".", copyright_holder = getOption("devtools.name",
  "<Author>"))

use_gpl3_license(pkg = ".")

use_dev_version(pkg = ".")

revdep_email(pkg = ".", date, version, author = getOption("devtools.name"),
  draft = TRUE, unsent = NULL, template = "revdep/email.md",
  only_problems = TRUE)

revdep_check_save_summary(pkg = ".")

revdep_check_print_problems(pkg = ".")

revdep_check(pkg = ".", recursive = FALSE, ignore = NULL,
  dependencies = c("Depends", "Imports", "Suggests", "LinkingTo"),
  skip = character(), libpath = getOption("devtools.revdep.libpath"),
  srcpath = libpath, bioconductor = FALSE, type = getOption("pkgType"),
  threads = getOption("Ncpus", 1), env_vars = NULL, check_dir = NULL,
  install_dir = NULL, quiet_check = TRUE)

revdep_check_resume(pkg = ".", ...)

revdep_check_reset(pkg = ".")
}
\description{
These functions are Deprecated in this release of devtools, they will be
marked as Defunct and removed in a future version.
}
\details{
Instead of the use_xyz functions from devtools use \link[usethis]{use_testthat}.

Instead of the revdep functions in devtools a better alternative is
\sQuote{revdepcheck::revdep_check()}.
}
\keyword{internal}
