% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test.R
\name{devtools-deprecated}
\alias{devtools-deprecated}
\alias{test_file}
\alias{test_coverage_file}
\title{Deprecated functions}
\usage{
test_file(file = find_active_file(), ...)

test_coverage_file(file = find_active_file(), ...)
}
\description{
These functions are Deprecated in this release of devtools, they will be
marked as Defunct and removed in a future version.
}
\keyword{internal}
