\name{document}
\alias{document}
\title{Use roxygen to make documentation.}
\usage{document(pkg)}

\description{
  Use roxygen to make documentation.
}
\keyword{programming}
\arguments{
  \item{pkg}{package description, can be path or package name.  See
\code{\link{as.package}} for more information}
}
