% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/document.R
\name{document}
\alias{document}
\title{Use roxygen to document a package.}
\usage{
document(pkg = ".", roclets = NULL, quiet = FALSE)
}
\arguments{
\item{pkg}{package description, can be path or package name.  See
\code{\link[=as.package]{as.package()}} for more information}

\item{roclets}{Character vector of roclet names to use with package.
This defaults to \code{NULL}, which will use the \code{roclets} fields in
the list provided in the \code{Roxygen} DESCRIPTION field. If none are
specified, defaults to \code{c("collate", "namespace", "rd")}.}

\item{quiet}{if \code{TRUE} suppresses output from this function.}
}
\description{
This function is a wrapper for the \code{\link[roxygen2:roxygenize]{roxygen2::roxygenize()}}
function from the roxygen2 package. See the documentation and vignettes of
that package to learn how to use roxygen.
}
\seealso{
\code{\link[roxygen2:roxygenize]{roxygen2::roxygenize()}},
\code{browseVignettes("roxygen2")}
}
