\name{document}
\alias{document}
\title{Use roxygen to make documentation.}
\usage{document(pkg, check=TRUE)}

\description{
  Use roxygen to make documentation.
}
\keyword{programming}
\arguments{
  \item{pkg}{package description, can be path or package name.  See
\code{\link{as.package}} for more information}
  \item{check}{if \code{TRUE} (the default), will automatically 
check documentation after running roxygen.}
}
