% Generated by roxygen2 (4.0.2): do not edit by hand
\name{document}
\alias{document}
\title{Use roxygen to document a package.}
\usage{
document(pkg = ".", clean = NULL, roclets = NULL, reload = TRUE)
}
\arguments{
\item{pkg}{package description, can be path or package name.  See
\code{\link{as.package}} for more information}

\item{clean,reload}{Deprecated.}

\item{roclets}{Character vector of roclet names to use with package.
This defaults to \code{NULL}, which will use the \code{roclets} fields in
the list provided in the \code{Roxygen} DESCRIPTION field. If none are
specified, defaults to \code{c("collate", "namespace", "rd")}.}
}
\description{
This function is a wrapper for the \code{\link[roxygen2]{roxygenize}()}
from the roxygen2 package. See the documentation and vignettes of
that package to learn how to use roxygen.
}
\seealso{
\code{\link[roxygen2]{roxygenize}},
  \code{browseVignettes("roxygen2")}
}

