\name{document}
\alias{document}
\title{Use roxygen to make documentation.}
\usage{
  document(pkg = NULL, clean = FALSE)
}
\arguments{
  \item{pkg}{package description, can be path or package
  name.  See \code{\link{as.package}} for more information}

  \item{clean}{if \code{TRUE} will automatically clear all
  roxygen caches and delete currently man contents to
  ensure that you have the freshest version of the
  documentation. check documentation after running
  roxygen.}
}
\description{
  Use roxygen to make documentation.
}
\keyword{programming}

