% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/doctor.R
\name{dr_github}
\alias{dr_github}
\title{Diagnose potential github issues}
\usage{
dr_github(path = ".")
}
\arguments{
\item{path}{Path to repository to check. Defaults to current working
directory}
}
\description{
Diagnose potential github issues
}
\examples{
\donttest{
dr_github()
}
}
\seealso{
Other doctors: \code{\link{dr_devtools}}
}

