% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/github.R
\name{github_pat}
\alias{github_pat}
\title{Retrieve Github personal access token.}
\usage{
github_pat(quiet = FALSE)
}
\description{
A github personal access token
Looks in env var \code{GITHUB_PAT}
}
\keyword{internal}
