% Generated by roxygen2 (4.0.2): do not edit by hand
\name{github_pull}
\alias{github_pull}
\alias{github_release}
\title{GitHub references}
\usage{
github_pull(pull)

github_release()
}
\arguments{
\item{pull}{The pull request to install}
}
\description{
Use as \code{ref} parameter to \code{\link{install_github}}.
Allows installing a specific pull request or the latest release.
}
\seealso{
\code{\link{install_github}}
}

