\name{has_devel}
\alias{has_devel}
\title{Check if you have a development environment installed.}

\description{
  Check if you have a development environment installed.
}

\details{
  Thanks to the suggestion on Simon Urbanek.
}
\value{TRUE if your development environment is correctly set up, otherwise
returns an error.}
\examples{has_devel()}
