\name{has_devel}
\alias{has_devel}
\title{Check if you have a development environment installed.}
\usage{
has_devel()
}
\value{
  TRUE if your development environment is correctly set up,
  otherwise returns an error.
}
\description{
  Thanks to the suggestion of Simon Urbanek.
}
\examples{
has_devel()
}

