% Generated by roxygen2 (4.0.2): do not edit by hand
\name{help}
\alias{?}
\alias{help}
\alias{shim_help}
\alias{shim_question}
\title{Drop-in replacements for help and ? functions}
\usage{
# help(topic, package = NULL, ...)

# ?e2
# e1?e2
}
\arguments{
\item{topic}{A name or character string specifying the help topic.}

\item{package}{A name or character string specifying the package in which
to search for the help topic. If NULL, seach all packages.}

\item{...}{Additional arguments to pass to \code{\link[utils]{help}}.}

\item{e1}{First argument to pass along to \code{utils::`?`}.}

\item{e2}{Second argument to pass along to \code{utils::`?`}.}
}
\description{
The \code{?} and \code{help} functions are replacements for functions of the
same name in the utils package. They are made available when a package is
loaded with \code{\link{load_all}}.
}
\details{
The \code{?} function is a replacement for \code{\link[utils]{?}} from the
utils package. It will search for help in devtools-loaded packages first,
then in regular packages.

The \code{help} function is a replacement for \code{\link[utils]{help}} from
the utils package. If \code{package} is not specified, it will search for
help in devtools-loaded packages first, then in regular packages. If
\code{package} is specified, then it will search for help in devtools-loaded
packages or regular packages, as appropriate.
}
\examples{
\dontrun{
# This would load devtools and look at the help for load_all, if currently
# in the devtools source directory.
load_all()
?load_all
help("load_all")
}

# To see the help pages for utils::help and utils::`?`:
help("help", "utils")
help("?", "utils")

\dontrun{
# Examples demonstrating the multiple ways of supplying arguments
# NB: you can't do pkg <- "ggplot2"; help("ggplot2", pkg)
shim_help(lm)
shim_help(lm, stats)
shim_help(lm, 'stats')
shim_help('lm')
shim_help('lm', stats)
shim_help('lm', 'stats')
shim_help(package = stats)
shim_help(package = 'stats')
topic <- "lm"
shim_help(topic)
shim_help(topic, stats)
shim_help(topic, 'stats')
}
}

