\name{install}
\alias{install}
\title{Install a package.}
\usage{install(pkg)}

\description{
  Install a package.
}

\details{
  Uses \code{R CMD install} to install the package.
}
\arguments{
  \item{pkg}{package description, can be path or package name.  See
\code{\link{as.package}} for more information}
}
