% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/install-cran.r
\name{install_cran}
\alias{install_cran}
\title{Attempts to install a package from CRAN.}
\usage{
install_cran(pkgs, repos = getOption("repos"), type = getOption("pkgType"),
  ..., quiet = FALSE)
}
\arguments{
\item{pkgs}{Character vector of packages to install.}
}
\description{
This function is vectorised on \code{pkgs} so you can install multiple
packages in a single command.
}
\examples{
\dontrun{
install_cran("ggplot2")
install_cran(c("httpuv", "shiny")
}
}
\seealso{
Other package installation: \code{\link{install_bioc}},
  \code{\link{install}}, \code{\link{uninstall}}
}

