% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/install.r
\name{install_deps}
\alias{install_deps}
\title{Install package dependencies if needed.}
\usage{
install_deps(pkg = ".", dependencies = NA, threads = getOption("Ncpus",
  1), ...)
}
\arguments{
\item{pkg}{package description, can be path or package name.  See
\code{\link{as.package}} for more information}

\item{dependencies}{\code{logical} indicating to also install uninstalled
packages which this \code{pkg} depends on/links to/suggests. See
argument \code{dependencies} of \code{\link{install.packages}}.}

\item{threads}{number of concurrent threads to use for installing
dependencies.
It defaults to the option \code{"Ncpus"} or \code{1} if unset.}

\item{...}{additional arguments passed to \code{\link{install.packages}}.}
}
\description{
Install package dependencies if needed.
}
\examples{
\dontrun{install_deps(".")}
}

