% Generated by roxygen2 (4.0.0): do not edit by hand
\name{install_deps}
\alias{install_deps}
\title{Install package dependencies}
\usage{
install_deps(pkg = ".", dependencies = NA, threads = getOption("Ncpus",
  1))
}
\arguments{
\item{pkg}{package description, can be path or package name.  See
\code{\link{as.package}} for more information}

\item{dependencies}{\code{logical} indicating to also install uninstalled
packages which this \code{pkg} depends on/links to/suggests. See
argument \code{dependencies} of \code{\link{install.packages}}.}

\item{threads}{number of concurrent threads to use for installing
dependencies.
It defaults to the option \code{"Ncpus"} or \code{1} if unset.}
}
\description{
Install package dependencies
}
\examples{
\dontrun{install_deps(".")}
}

