% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/install.r
\name{install_dev_deps}
\alias{install_dev_deps}
\title{Install package dependencies if needed.}
\usage{
install_dev_deps(pkg = ".", ...)
}
\arguments{
\item{pkg}{package description, can be path or package name.  See
\code{\link{as.package}} for more information}

\item{...}{additional arguments passed to \code{\link[utils]{install.packages}}.}
}
\description{
Install package dependencies if needed.
}
\examples{
\dontrun{install_deps(".")}
}
