% Generated by roxygen2 (4.0.1): do not edit by hand
\name{install_github_enterprise}
\alias{install_github_enterprise}
\title{Convenience wrapper for \code{\link{install_github}}.}
\usage{
install_github_enterprise(repo, username = NULL, ref = "master",
  subdir = NULL, auth_token = github_pat(),
  github_url = devtools_git_enterprise(), ..., dependencies = TRUE)
}
\arguments{
\item{repo}{Repository address in the format
\code{[username/]repo[/subdir][@ref|#pull]}. Alternatively, you can
specify \code{username}, \code{subdir}, \code{ref} or \code{pull} using the
respective parameters (see below); if both is specified, the values in
\code{repo} take precedence.}

\item{username}{User name. Deprecate: please include username in the
\code{repo}}

\item{ref}{Desired git reference. Could be a commit, tag, or branch
name, or a call to \code{\link{github_pull}}. Defaults to \code{"master"}.}

\item{subdir}{subdirectory within repo that contains the R package.}

\item{auth_token}{To install from a private repo, generate a personal
access token (PAT) in \url{https://github.com/settings/applications} and
supply to this argument. This is safer than using a password because
you can easily delete a PAT without affecting any others. Defaults to
the \code{GITHUB_PAT} environment variable.}

\item{github_url}{Defaults to NULL, so the default archive URL is served
from the GitHub API. You can set it to your custom Enterprise GitHub URL.}

\item{...}{Other arguments passed on to \code{\link{install}}.}

\item{dependencies}{By default, installs all dependencies so that you can
build vignettes and use all functionality of the package.}
}
\description{
This function allows you to install a package built
with \code{devtools} from a repo with a custom URL.
}
\examples{
\dontrun{
# To install from a private repo, use auth_token as described
# at ?install_github and either set the github_url parameter
# or let the ?devtools_git_enterprise helper retrieve it from
# the GITHUB_URL environment variable if it is set.
# Best practice is the latter.
install_github_enterprise("username/packagename", github_url = "https://github.scm.xyz.com")

}
}
\seealso{
Other package installation: \code{\link{install_bitbucket}};
  \code{\link{install_github}};
  \code{\link{install_gitorious}};
  \code{\link{install_git}}; \code{\link{install_svn}};
  \code{\link{install_url}}; \code{\link{install_version}};
  \code{\link{install}}
}

