% Generated by roxygen2 (4.0.2): do not edit by hand
\name{install_gitorious}
\alias{install_gitorious}
\title{Attempts to install a package directly from gitorious.}
\usage{
install_gitorious(repo, ref = "master", subdir = NULL, ...)
}
\arguments{
\item{repo}{Repository address in the format
\code{username/repo[/subdir][@ref|#pull]}. Alternatively, you can
specify \code{subdir} and/or \code{ref} using the respective parameters
(see below); if both is specified, the values in \code{repo} take
precedence.}

\item{ref}{Desired git reference. Could be a commit, tag, or branch
name, or a call to \code{\link{github_pull}}. Defaults to \code{"master"}.}

\item{subdir}{subdirectory within repo that contains the R package.}

\item{...}{Other arguments passed on to \code{\link{install}}.}
}
\description{
This function is vectorised so you can install multiple packages in
a single command.
}
\examples{
\dontrun{
install_gitorious("r-mpc-package/r-mpc-package")
}
}
\seealso{
Other package installation: \code{\link{install_bitbucket}};
  \code{\link{install_github}}; \code{\link{install_git}};
  \code{\link{install_svn}}; \code{\link{install_url}};
  \code{\link{install_version}}; \code{\link{install}}
}

