% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/install-svn.r
\name{install_svn}
\alias{install_svn}
\title{Install a package from a SVN repository}
\usage{
install_svn(url, subdir = NULL, args = character(0), ..., revision = NULL,
  quiet = FALSE)
}
\arguments{
\item{url}{Location of package. The url should point to a public or
private repository.}

\item{subdir}{A sub-directory withing a svn repository that may contain the
package we are interested in installing.}

\item{args}{A character vector providing extra arguments to pass on to}

\item{...}{Other arguments passed on to \code{\link{install}}}

\item{revision}{svn revision, if omitted updates to latest}

\item{quiet}{if \code{TRUE} suppresses output from this function.}
}
\description{
This function requires \code{svn} to be installed on your system in order to
be used.
}
\details{
It is vectorised so you can install multiple packages with
a single command.
}
\examples{
\dontrun{
install_svn("https://github.com/hadley/stringr")
install_svn("https://github.com/hadley/httr")
}
}
\seealso{
Other package installation: \code{\link{install_bioc}},
  \code{\link{install_bitbucket}},
  \code{\link{install_cran}}, \code{\link{install_github}},
  \code{\link{install_gitlab}}, \code{\link{install_git}},
  \code{\link{install_url}}, \code{\link{install_version}},
  \code{\link{install}}, \code{\link{uninstall}}
}
