% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/install-version.r
\name{install_version}
\alias{install_version}
\title{Install specified version of a CRAN package.}
\usage{
install_version(package, version = NULL, repos = getOption("repos"),
  type = getOption("pkgType"), ..., quiet = FALSE)
}
\arguments{
\item{package}{package name}

\item{version}{If the specified version is NULL or the same as the most
recent version of the package, this function simply calls
\code{\link{install}}. Otherwise, it looks at the list of
archived source tarballs and tries to install an older version instead.}

\item{repos}{
    character vector, the base URL(s) of the repositories
    to use, e.g., the URL of a CRAN mirror such as
    \code{"https://cloud.r-project.org"}.  For more details on
    supported URL schemes see \code{\link{url}}.

    Can be \code{NULL} to install from local files, directories or URLs:
    this will be inferred by extension from \code{pkgs} if of length one.
  }

\item{type}{character, indicating the type of package to download and
    install.  Will be \code{"source"} except on Windows and some macOS
    builds: see the section on \sQuote{Binary packages} for those.
  }

\item{...}{Other arguments passed on to \code{\link{install.packages}}.}

\item{quiet}{
    logical: if true, reduce the amount of output.
  }
}
\description{
If you are installing an package that contains compiled code, you will
need to have an R development environment installed.  You can check
if you do by running \code{\link[pkgbuild]{has_build_tools}}.
}
\seealso{
Other package installation: \code{\link{install_bioc}},
  \code{\link{install_bitbucket}},
  \code{\link{install_cran}}, \code{\link{install_github}},
  \code{\link{install_gitlab}}, \code{\link{install_git}},
  \code{\link{install_svn}}, \code{\link{install_url}},
  \code{\link{install}}, \code{\link{uninstall}}
}
\author{
Jeremy Stephens
}
