% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/lint.r
\name{lint}
\alias{lint}
\title{Lint all source files in a package.}
\usage{
lint(pkg = ".", cache = TRUE, ...)
}
\arguments{
\item{pkg}{package description, can be path or package name. See
\code{\link{as.package}} for more information}

\item{cache}{store the lint results so repeated lints of the same content
use the previous results.}

\item{...}{additional arguments passed to \code{\link[lintr]{lint_package}}}
}
\description{
The default lintings correspond to the style guide at
\url{http://r-pkgs.had.co.nz/r.html#style}, however it is possible to
override any or all of them using the \code{linters} parameter.
}
\details{
The lintr cache is by default stored in \code{~/.R/lintr_cache/} (this can
be configured by setting \code{options(lintr.cache_directory)}).  It can be
cleared by calling \code{\link[lintr]{clear_cache}}.
}
\seealso{
\code{\link[lintr]{lint_package}}, \code{\link[lintr]{lint}}
}

