\name{load_all}
\alias{load_all}
\title{Load complete package.}
\usage{load_all(pkg, reset=FALSE)}

\description{
  Load complete package.
}
\keyword{programming}
\arguments{
  \item{pkg}{package description, can be path or package name.  See
\code{\link{as.package}} for more information}
  \item{reset}{clear package environment and reset file cache before loading
any pieces of the package.}
}
