% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/load-data.r
\name{load_data}
\alias{load_data}
\title{Load data.}
\usage{
load_data(pkg = ".")
}
\arguments{
\item{pkg}{package description, can be path or package name.  See
\code{\link{as.package}} for more information}
}
\description{
Loads all \code{.RData} files in the data subdirectory.
}
\keyword{programming}

