\name{load_deps}
\alias{load_deps}
\title{Load dependencies.}
\usage{
  load_deps(pkg = NULL)
}
\arguments{
  \item{pkg}{package description, can be path or package
  name.  See \code{\link{as.package}} for more information}
}
\description{
  Load all package dependencies as described in
  \code{DESCRIPTION}.
}
\keyword{programming}

