% Generated by roxygen2 (4.0.2): do not edit by hand
\name{loaded_packages}
\alias{loaded_packages}
\title{Return a vector of names of attached packages}
\usage{
loaded_packages()
}
\value{
A data frame with columns package and path, giving the name of
  each package and the path it was loaded from.
}
\description{
Return a vector of names of attached packages
}
\keyword{internal}

