% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/path.r
\name{on_path}
\alias{on_path}
\title{Test if an object is on the path.}
\usage{
on_path(...)
}
\arguments{
\item{...}{Strings indicating the executables to check for on the path.}
}
\description{
Test if an object is on the path.
}
\examples{
on_path("R")
on_path("gcc")
on_path("foo", "bar")  # FALSE in most cases
withr::with_path(tempdir(), on_path("gcc"))
}
\seealso{
Other path: \code{\link{path}}
}
\keyword{internal}
