\name{print.envlist}
\alias{print.envlist}
\title{Print an \code{envlist}}
\usage{
\method{print}{envlist}(x, name = TRUE, path = FALSE, ...)
}
\arguments{
  \item{x}{An \code{envlist} object to print.}

  \item{name}{If \code{TRUE} (the default), print the
  \code{name} attribute of each environment.}

  \item{path}{If \code{TRUE}, print the \code{path}
  attribute of each environment.}

  \item{...}{Other arguments to be passed to \code{print}.}
}
\description{
  Print an \code{envlist}
}
\keyword{internal}

