% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/infrastructure-git.R, R/infrastructure.R,
%   R/pkgbuild.R, R/pkgload.R
\docType{import}
\name{reexports}
\alias{reexports}
\alias{use_git}
\alias{reexports}
\alias{use_github}
\alias{reexports}
\alias{use_git_hook}
\alias{reexports}
\alias{use_github_links}
\alias{reexports}
\alias{use_testthat}
\alias{reexports}
\alias{use_test}
\alias{reexports}
\alias{use_rstudio}
\alias{reexports}
\alias{use_vignette}
\alias{reexports}
\alias{use_rcpp}
\alias{reexports}
\alias{use_travis}
\alias{reexports}
\alias{use_coverage}
\alias{reexports}
\alias{use_appveyor}
\alias{reexports}
\alias{use_package_doc}
\alias{reexports}
\alias{use_package}
\alias{reexports}
\alias{use_data}
\alias{reexports}
\alias{use_data_raw}
\alias{reexports}
\alias{use_build_ignore}
\alias{reexports}
\alias{use_readme_rmd}
\alias{reexports}
\alias{use_readme_md}
\alias{reexports}
\alias{use_news_md}
\alias{reexports}
\alias{use_revdep}
\alias{reexports}
\alias{use_cran_comments}
\alias{reexports}
\alias{use_code_of_conduct}
\alias{reexports}
\alias{use_cran_badge}
\alias{reexports}
\alias{use_mit_license}
\alias{reexports}
\alias{use_gpl3_license}
\alias{reexports}
\alias{use_dev_version}
\alias{reexports}
\alias{build}
\alias{reexports}
\alias{with_debug}
\alias{reexports}
\alias{load_all}
\alias{reexports}
\alias{unload}
\title{Objects exported from other packages}
\keyword{internal}
\description{
These objects are imported from other packages. Follow the links
below to see their documentation.

\describe{
  \item{pkgbuild}{\code{\link[pkgbuild]{build}}, \code{\link[pkgbuild]{with_debug}}}

  \item{pkgload}{\code{\link[pkgload]{load_all}}, \code{\link[pkgload]{unload}}}

  \item{usethis}{\code{\link[usethis]{use_git}}, \code{\link[usethis]{use_github}}, \code{\link[usethis]{use_git_hook}}, \code{\link[usethis]{use_github_links}}, \code{\link[usethis]{use_testthat}}, \code{\link[usethis]{use_test}}, \code{\link[usethis]{use_rstudio}}, \code{\link[usethis]{use_vignette}}, \code{\link[usethis]{use_rcpp}}, \code{\link[usethis]{use_travis}}, \code{\link[usethis]{use_coverage}}, \code{\link[usethis]{use_appveyor}}, \code{\link[usethis]{use_package_doc}}, \code{\link[usethis]{use_package}}, \code{\link[usethis]{use_data}}, \code{\link[usethis]{use_data_raw}}, \code{\link[usethis]{use_build_ignore}}, \code{\link[usethis]{use_readme_rmd}}, \code{\link[usethis]{use_readme_md}}, \code{\link[usethis]{use_news_md}}, \code{\link[usethis]{use_revdep}}, \code{\link[usethis]{use_cran_comments}}, \code{\link[usethis]{use_code_of_conduct}}, \code{\link[usethis]{use_cran_badge}}, \code{\link[usethis]{use_mit_license}}, \code{\link[usethis]{use_gpl3_license}}, \code{\link[usethis]{use_dev_version}}}
}}

