% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/remotes.R
\docType{import}
\name{reexports}
\alias{dev_package_deps}
\alias{github_pull}
\alias{github_release}
\alias{install_bitbucket}
\alias{install_git}
\alias{install_github}
\alias{install_local}
\alias{install_svn}
\alias{install_url}
\alias{install_version}
\alias{package_deps}
\alias{reexports}
\alias{update_packages}
\title{Objects exported from other packages}
\description{
These objects are imported from other packages. Follow the links
below to see their documentation.

\describe{
  \item{remotes}{\code{\link[remotes]{update_packages}}, \code{\link[remotes]{package_deps}}, \code{\link[remotes]{dev_package_deps}}, \code{\link[remotes]{install_bitbucket}}, \code{\link[remotes]{install_git}}, \code{\link[remotes]{install_github}}, \code{\link[remotes]{github_pull}}, \code{\link[remotes]{github_release}}, \code{\link[remotes]{install_local}}, \code{\link[remotes]{install_svn}}, \code{\link[remotes]{install_url}}, \code{\link[remotes]{install_version}}}
}}
\keyword{internal}

