% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inst.R, R/package-deps.R, R/pkgbuild.R,
%   R/pkgload.R, R/remotes.R, R/session-info.R
\docType{import}
\name{reexports}
\alias{reexports}
\alias{inst}
\alias{parse_deps}
\alias{check_dep_version}
\alias{with_debug}
\alias{clean_dll}
\alias{has_devel}
\alias{find_rtools}
\alias{unload}
\alias{install_bioc}
\alias{install_bitbucket}
\alias{install_cran}
\alias{install_dev}
\alias{install_git}
\alias{install_github}
\alias{github_pull}
\alias{github_release}
\alias{install_gitlab}
\alias{install_local}
\alias{install_svn}
\alias{install_url}
\alias{install_version}
\alias{update_packages}
\alias{dev_package_deps}
\alias{session_info}
\alias{package_info}
\title{Objects exported from other packages}
\keyword{internal}
\description{
These objects are imported from other packages. Follow the links
below to see their documentation.

\describe{
  \item{pkgbuild}{\code{\link[pkgbuild]{with_debug}}, \code{\link[pkgbuild]{clean_dll}}, \code{\link[pkgbuild]{has_devel}}, \code{\link[pkgbuild]{find_rtools}}}

  \item{pkgload}{\code{\link[pkgload]{inst}}, \code{\link[pkgload]{parse_deps}}, \code{\link[pkgload]{check_dep_version}}, \code{\link[pkgload]{unload}}}

  \item{remotes}{\code{\link[remotes]{install_bioc}}, \code{\link[remotes]{install_bitbucket}}, \code{\link[remotes]{install_cran}}, \code{\link[remotes]{install_dev}}, \code{\link[remotes]{install_git}}, \code{\link[remotes]{install_github}}, \code{\link[remotes]{github_pull}}, \code{\link[remotes]{github_release}}, \code{\link[remotes]{install_gitlab}}, \code{\link[remotes]{install_local}}, \code{\link[remotes]{install_svn}}, \code{\link[remotes]{install_url}}, \code{\link[remotes]{install_version}}, \code{\link[remotes]{update_packages}}, \code{\link[remotes]{dev_package_deps}}}

  \item{sessioninfo}{\code{\link[sessioninfo]{session_info}}, \code{\link[sessioninfo]{package_info}}}
}}

