% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package-deps.r, R/pkgbuild.R, R/pkgload.R
\docType{import}
\name{reexports}
\alias{reexports}
\alias{parse_deps}
\alias{reexports}
\alias{check_dep_version}
\alias{reexports}
\alias{with_debug}
\alias{reexports}
\alias{clean_dll}
\alias{reexports}
\alias{unload}
\title{Objects exported from other packages}
\keyword{internal}
\description{
These objects are imported from other packages. Follow the links
below to see their documentation.

\describe{
  \item{pkgbuild}{\code{\link[pkgbuild]{with_debug}}, \code{\link[pkgbuild]{clean_dll}}}

  \item{pkgload}{\code{\link[pkgload]{parse_deps}}, \code{\link[pkgload]{check_dep_version}}, \code{\link[pkgload]{unload}}}
}}

