% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/release.R
\name{release}
\alias{release}
\title{Release package to CRAN}
\usage{
release(pkg = ".", check = FALSE, args = NULL)
}
\arguments{
\item{pkg}{The package to use, can be a file path to the package or a
package object.  See \code{\link[=as.package]{as.package()}} for more information.}

\item{check}{if \code{TRUE}, run checking, otherwise omit it.  This
is useful if you've just checked your package and you're ready to
release it.}

\item{args}{An optional character vector of additional command
line arguments to be passed to \verb{R CMD build}.}
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#deprecated}{\figure{lifecycle-deprecated.svg}{options: alt='[Deprecated]'}}}{\strong{[Deprecated]}}

\code{release()} is deprecated in favour of \code{\link[usethis:use_release_issue]{usethis::use_release_issue()}}.
We no longer feel confident recommending \code{release()} because we don't use it
ourselves, so there's no guarantee that it will track best practices as
they evolve over time.

If you want to programmatical submit to CRAN, you can continue to use
\code{\link[=submit_cran]{submit_cran()}}.
}
\seealso{
\code{\link[usethis:use_release_issue]{usethis::use_release_issue()}} to create a checklist of release
tasks that you can use in addition to or in place of \code{release}.
}
