% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check-devtools.R
\name{release_checks}
\alias{release_checks}
\title{Custom devtools release checks}
\usage{
release_checks(pkg = ".", built_path = NULL)
}
\arguments{
\item{pkg}{The package to use, can be a file path to the package or a
package object.  See \code{\link[=as.package]{as.package()}} for more information.}
}
\description{
This function performs additional checks prior to release. It is called
automatically by \code{\link[=release]{release()}}.
}
\keyword{internal}
