% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reload.R
\name{reload}
\alias{reload}
\title{Unload and reload package}
\usage{
reload(pkg = ".", quiet = FALSE)
}
\arguments{
\item{pkg}{The package to use, can be a file path to the package or a
package object.  See \code{\link[=as.package]{as.package()}} for more information.}

\item{quiet}{if \code{TRUE} suppresses output from this function.}
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#deprecated}{\figure{lifecycle-deprecated.svg}{options: alt='[Deprecated]'}}}{\strong{[Deprecated]}}

\code{reload()} is deprecated because we no longer use or recommend this
workflow. Instead, we recommend \code{\link[=load_all]{load_all()}} to load a package for
interactive development.
}
\seealso{
\code{\link[=load_all]{load_all()}} to load a package for interactive development.
}
\keyword{internal}
