\name{reload}
\alias{reload}
\title{Detach and reload package.}
\usage{reload(pkg)}

\description{
  Detach and reload package.
}

\details{
  See the caveats in \code{\link{detach}} for the many
  reasons why this might not work. If in doubt, quit R and
  restart.
}
\arguments{
  \item{pkg}{package description, can be path or package name.  See
\code{\link{as.package}} for more information}
}
