% Generated by roxygen2 (4.0.2): do not edit by hand
\name{revdep}
\alias{revdep}
\alias{revdep_maintainers}
\title{Reverse dependency tools.}
\usage{
revdep(pkg, dependencies = c("Depends", "Imports", "Suggests", "LinkingTo"),
  recursive = FALSE, ignore = NULL)

revdep_maintainers(pkg = ".")
}
\arguments{
\item{pkg}{Package name. This is unlike most devtools packages which
take a path because you might want to determine dependencies for a package
that you don't have installed. If omitted, defaults to the name of the
current package.}

\item{dependencies}{a character vector listing the types of
    dependencies, a subset of
    \code{c("Depends", "Imports", "LinkingTo", "Suggests", "Enhances")}.
    Character string \code{"all"} is shorthand for that vector.}

\item{recursive}{logical: should reverse dependencies of reverse
    dependencies (and so on) be included?}

\item{ignore}{A character vector of package names to ignore. These packages
will not appear in returned vector. This is used in
\code{\link{revdep_check}} to avoid packages with installation problems
or extremely long check times.}
}
\description{
Tools to check and notify maintainers of all CRAN and bioconductor
packages that depend on the specified package.
}
\details{
The first run in a session will be time-consuming because it must download
all package metadata from CRAN and bioconductor. Subsequent runs will
be faster.
}
\examples{
\dontrun{
revdep("ggplot2")

revdep("ggplot2", ignore = c("xkcd", "zoo"))
}
}
\seealso{
\code{\link{revdep_check}()} to run R CMD check on all reverse
  dependencies.
}

