% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/r-hub.R
\name{rhub_check}
\alias{rhub_check}
\alias{check_rhub}
\title{Run CRAN checks for package on R-hub}
\usage{
rhub_check(
  gh_url = NULL,
  platforms = c("ubuntu-release", "windows", "macos"),
  r_versions = NULL,
  branch = NULL
)

check_rhub()
}
\arguments{
\item{gh_url}{GitHub URL of a package to check, or \code{NULL} to check
the package in the current directory.}

\item{platforms}{Platforms to use, a character vector.
Defaults to most common platforms. Use NULL to select from
a list in interactive sessions. See \code{\link[rhub:rhub_platforms]{rhub::rhub_platforms()}}.}

\item{r_versions}{Which R version(s) to use for the platforms that
supports multiple R versions. This arguemnt is not implemented yet.}

\item{branch}{Branch to use to run R-hub. Defaults to the current
branch if \code{gh_url} is \code{NULL}. Otherwise defaults to \code{"main"}. Note that
this branch also need to include the \code{rhub.yaml} workflow file.}
}
\description{
It runs an R-hub GitHub action on the package.
Used to check if the package passes CRAN checks on
different platforms and R versions.
}
\section{GitHub and R-hub}{

To use R-hub you must first setup a GitHub repository.
Use \code{\link[usethis:use_github]{usethis::use_github()}} to do that. Then setup your
R-hub GH Action with \code{\link[rhub:rhub_setup]{rhub::rhub_setup()}}.
}

\seealso{
Other build functions: 
\code{\link{check_mac_release}()},
\code{\link{check_win}()}
}
\concept{build functions}
