% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/session-info.r
\name{session_info}
\alias{session_info}
\title{Print session information}
\usage{
session_info(include_base = FALSE)
}
\arguments{
\item{include_base}{Include base packages in summary? By default this is
false since base packages should always match the R version.}
}
\description{
This is \code{\link{sessionInfo}()} re-written from scratch to both exclude
data that's rarely useful (e.g., the full collate string or base packages
loaded) and include stuff you'd like to know (e.g., where a package was
installed from).
}

