% Generated by roxygen2 (4.0.2): do not edit by hand
\name{session_info}
\alias{session_info}
\title{Print session information}
\usage{
session_info()
}
\description{
This is \code{\link{sessionInfo}()} re-written from scratch to both exclude
data that's rarely useful (e.g., the full collate string or base packages
loaded) and include stuff you'd like to know (e.g., where a package was
installed from).
}

