\name{set_path}
\alias{set_path}
\title{set the path from a vector of individual paths}
\usage{
  set_path(path)
}
\arguments{
  \item{path}{character vector of locations}
}
\description{
  set the path from a vector of individual paths
}
\note{
  This only works for the current session.  The path is not
  set permanantly.
}

