% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/show-news.R
\name{show_news}
\alias{show_news}
\title{Show package news}
\usage{
show_news(pkg = ".", latest = TRUE, ...)
}
\arguments{
\item{pkg}{The package to use, can be a file path to the package or a
package object.  See \code{\link[=as.package]{as.package()}} for more information.}

\item{latest}{if \code{TRUE}, only show the news for the most recent
version.}

\item{...}{other arguments passed on to \code{news}}
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#deprecated}{\figure{lifecycle-deprecated.svg}{options: alt='[Deprecated]'}}}{\strong{[Deprecated]}}

\code{show_news()} is deprecated because we no longer use or recommend this
workflow.
}
\keyword{internal}
