% Generated by roxygen2 (4.0.0): do not edit by hand
\name{show_news}
\alias{show_news}
\title{Show package news}
\usage{
show_news(pkg = ".", latest = TRUE, ...)
}
\arguments{
\item{pkg}{package description, can be path or package name.  See
\code{\link{as.package}} for more information}

\item{latest}{if \code{TRUE}, only show the news for the most recent
version.}

\item{...}{other arguments passed on to \code{news}}
}
\description{
Show package news
}

