\name{show_rd}
\alias{show_rd}
\title{Show an Rd file in a package.}
\usage{
  show_rd(pkg = NULL, file)
}
\arguments{
  \item{pkg}{package description, can be path or package
  name.  See \code{\link{as.package}} for more information}

  \item{file}{name of Rd file to open.  Can optionally omit
  Rd extension.}
}
\description{
  Show an Rd file in a package.
}

