% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spell-check.R
\name{spell_check}
\alias{spell_check}
\title{Spell checking}
\usage{
spell_check(pkg = ".", ignore = character(), dict = "en_US")
}
\arguments{
\item{pkg}{package description, can be path or package name.  See
\code{\link{as.package}} for more information}

\item{ignore}{character vector with words to ignore. See
\code{\link[hunspell:hunspell]{hunspell}} for more information}

\item{dict}{a dictionary object or language string. See
\code{\link[hunspell:hunspell]{hunspell}} for more information}
}
\description{
Runs a spell check on text fields in the package description file and
manual pages. Hunspell includes dictionaries for \code{en_US} and \code{en_GB}
by default. Other languages require installation of a custom dictionary, see
the \href{https://cran.r-project.org/package=hunspell/vignettes/intro.html#system_dictionaries}{hunspell vignette}
for details.
}
