% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spell-check.R
\name{spell_check}
\alias{spell_check}
\title{Spell checking}
\usage{
spell_check(pkg = ".", vignettes = TRUE, lang = "en_US")
}
\arguments{
\item{pkg}{package description, can be path or package name.  See
\code{\link{as.package}} for more information}

\item{vignettes}{include vignettes in the spell check; passed to
\link[spelling:spell_check_package]{spelling::spell_check_package}}

\item{lang}{a dictionary object or language string; passed to
\link[spelling:spell_check_package]{spelling::spell_check_package}}
}
\description{
Runs a spell check on text fields in the package description file, manual
pages, and optionally vignettes. Wraps the \link[spelling:spell_check_package]{spelling}
package.
manual pages. Hunspell includes dictionaries for \code{en_US} and \code{en_GB}
by default. Other languages require installation of a custom dictionary, see
the \href{https://cran.r-project.org/package=hunspell/vignettes/intro.html#system_dictionaries}{hunspell vignette}
for details.
}
