% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/release.r
\name{submit_cran}
\alias{submit_cran}
\title{Submit a package to CRAN.}
\usage{
submit_cran(pkg = ".", args = NULL)
}
\arguments{
\item{pkg}{package description, can be path or package name.  See
\code{\link[=as.package]{as.package()}} for more information}

\item{args}{An optional character vector of additional command
line arguments to be passed to \code{R CMD build}.}
}
\description{
This uses the new CRAN web-form submission process. After submission, you
will receive an email asking you to confirm submission - this is used
to check that the package is submitted by the maintainer.
}
\details{
It's recommended that you use \code{\link[=release]{release()}} rather than this
function as it performs more checks prior to submission.
}
\keyword{internal}
