% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/release.r
\name{submit_cran}
\alias{submit_cran}
\title{Submit a package to CRAN.}
\usage{
submit_cran(pkg = ".")
}
\arguments{
\item{pkg}{package description, can be path or package name.  See
\code{\link{as.package}} for more information}
}
\description{
This uses the new CRAN web-form submission process. After submission, you
will receive an email asking you to confirm submission - this is used
to check that the package is submitted by the maintainer.
}
\details{
It's recommend that you use \code{\link{release}()} rather than this
function as it performs more checks prior to submission.
}
\keyword{internal}

