% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/release.R
\name{submit_cran}
\alias{submit_cran}
\title{Submit a package to CRAN.}
\usage{
submit_cran(pkg = ".", args = NULL)
}
\arguments{
\item{pkg}{The package to use, can be a file path to the package or a
package object.  See \code{\link[=as.package]{as.package()}} for more information.}

\item{args}{An optional character vector of additional command
line arguments to be passed to \verb{R CMD build}.}
}
\description{
This uses the new CRAN web-form submission process. After submission, you
will receive an email asking you to confirm submission - this is used
to check that the package is submitted by the maintainer.
}
\details{
It's recommended that you use \code{\link[=release]{release()}} rather than this
function as it performs more checks prior to submission.
Note however that the function is still part of the task list
generated by \code{\link[usethis:use_release_issue]{usethis::use_release_issue()}}, since those tasks more or less
reflect the checks done by \code{\link[=release]{release()}}.
}
\keyword{internal}
