% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/system.r
\name{system_check}
\alias{system_check}
\title{Run a system command and check if it succeeds.}
\usage{
system_check(cmd, args = character(), env = character(), quiet = FALSE,
  ...)
}
\arguments{
\item{cmd}{the command to run.}

\item{args}{a vector of command arguments.}

\item{env}{a named character vector of environment variables.  Will be quoted}

\item{quiet}{if \code{FALSE}, the command to be run will be echoed.}

\item{...}{additional arguments passed to \code{\link[base]{system}}}
}
\value{
\code{TRUE} if the command succeeds, an error will be thrown if the
command fails.
}
\description{
Run a system command and check if it succeeds.
}

