% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test.r
\name{test}
\alias{test}
\alias{test_coverage}
\alias{uses_testthat}
\title{Execute all \pkg{test_that} tests in a package.}
\usage{
test(pkg = ".", filter = NULL, ...)

test_coverage(pkg = ".", ...)

uses_testthat(pkg = ".")
}
\arguments{
\item{pkg}{package description, can be path or package name. See
\code{\link[=as.package]{as.package()}} for more information}

\item{filter}{If not \code{NULL}, only tests with file names matching this
regular expression will be executed.  Matching will take on the file
name after it has been stripped of \code{"test-"} and \code{".R"}.}

\item{...}{additional arguments passed to \code{\link[testthat:test_dir]{testthat::test_dir()}} and
\code{\link[covr:package_coverage]{covr::package_coverage()}}}
}
\description{
\code{test()} is a shortcut for \code{\link[testthat:test_dir]{testthat::test_dir()}}.
\code{test_coverage()} is a shortcut for \code{\link[covr:package_coverage]{covr::package_coverage()}}.
}
