% Generated by roxygen2 (4.0.2): do not edit by hand
\name{use_data_raw}
\alias{use_data_raw}
\title{Use \code{data-raw} to compute package datasets.}
\usage{
use_data_raw(pkg = ".")
}
\arguments{
\item{pkg}{Package where to create \code{data-raw}. Defaults to package in
working directory.}
}
\description{
Use \code{data-raw} to compute package datasets.
}
\seealso{
Other infrastructure: \code{\link{add_build_ignore}},
  \code{\link{use_build_ignore}};
  \code{\link{add_rstudio_project}},
  \code{\link{add_test_infrastructure}},
  \code{\link{add_travis}}, \code{\link{infrastructure}},
  \code{\link{use_package_doc}}, \code{\link{use_rcpp}},
  \code{\link{use_rstudio}}, \code{\link{use_testthat}},
  \code{\link{use_travis}}, \code{\link{use_vignette}};
  \code{\link{use_data}}; \code{\link{use_package}}
}

