% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/infrastructure-git.R
\name{use_github}
\alias{use_github}
\title{Connect a local repo with github.}
\usage{
use_github(auth_token = github_pat(), private = FALSE, pkg = ".")
}
\arguments{
\item{auth_token}{To install from a private repo, generate a personal
access token (PAT) in \url{https://github.com/settings/applications} and
supply to this argument. This is safer than using a password because
you can easily delete a PAT without affecting any others. Defaults to
the \code{GITHUB_PAT} environment variable.}

\item{private}{If \code{TRUE}, creates a private repository.}

\item{pkg}{Path to package. See \code{\link{as.package}} for more
information.}
}
\description{
If the current repo does not use git, calls \code{use_git} automatically.
}
\seealso{
Other git infrastructure: \code{\link{use_git_hook}};
  \code{\link{use_git}}
}
\keyword{internal}

