% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/infrastructure-git.R
\name{use_github_links}
\alias{use_github_links}
\title{Add GitHub links to DESCRIPTION.}
\usage{
use_github_links(pkg = ".")
}
\arguments{
\item{pkg}{Path to package. See \code{\link{as.package}} for more
information.}
}
\description{
Populates the URL and BugReports fields of DESCRIPTION with
\code{https://github.com/USERNAME/PKG} AND
\code{https://github.com/USERNAME/PKG/issues}, respectively. If package does
not already use GitHub (and therefore git), nothing will be done.
}
\seealso{
Other git infrastructure: \code{\link{use_git_hook}};
  \code{\link{use_github}}; \code{\link{use_git}}
}
\keyword{internal}

