% Generated by roxygen2 (4.0.2): do not edit by hand
\name{use_package}
\alias{use_package}
\title{Use specified package.}
\usage{
use_package(package, type = "Imports", pkg = ".")
}
\arguments{
\item{package}{Name of package to depend on.}

\item{type}{Type of dependency: must be one of "Imports", "Suggests",
"Depends", "Suggests", "Enhances", or "LinkingTo" (or unique abbreviation)}

\item{pkg}{package description, can be path or package name. See
\code{\link{as.package}} for more information.}
}
\description{
This adds a dependency to DESCRIPTION and offers a little advice
about how to best use it.
}
\examples{
\dontrun{
use_package("ggplot2")
use_package("dplyr", "suggests")

}
}
\seealso{
Other infrastructure: \code{\link{add_build_ignore}},
  \code{\link{use_build_ignore}};
  \code{\link{add_rstudio_project}},
  \code{\link{add_test_infrastructure}},
  \code{\link{add_travis}}, \code{\link{infrastructure}},
  \code{\link{use_package_doc}}, \code{\link{use_rcpp}},
  \code{\link{use_rstudio}}, \code{\link{use_testthat}},
  \code{\link{use_travis}}, \code{\link{use_vignette}};
  \code{\link{use_data_raw}}; \code{\link{use_data}}
}

