% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/infrastructure.R
\name{use_readme_rmd}
\alias{use_readme_rmd}
\title{Use README.Rmd}
\usage{
use_readme_rmd(pkg = ".")
}
\arguments{
\item{pkg}{package description, can be path or package name.  See
\code{\link{as.package}} for more information}

\item{hook}{Hook name. One of "pre-commit", "prepare-commit-msg",
"commit-msg", "post-commit", "applypatch-msg", "pre-applypatch",
"post-applypatch", "pre-rebase", "post-rewrite", "post-checkout",
"post-merge", "pre-push", "pre-auto-gc".}

\item{script}{Text of script to run}
}
\description{
This creates `README.Rmd` from template and adds to \code{.Rbuildignore}.
}
\seealso{
Other infrastructure: \code{\link{infrastructure}},
  \code{\link{use_build_ignore}},
  \code{\link{use_data_raw}}, \code{\link{use_data}},
  \code{\link{use_news_md}}, \code{\link{use_package}}
}
\keyword{internal}

