% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/infrastructure.R
\name{use_readme_rmd}
\alias{use_readme_md}
\alias{use_readme_rmd}
\title{Create README files.}
\usage{
use_readme_rmd(pkg = ".")

use_readme_md(pkg = ".")
}
\arguments{
\item{pkg}{package description, can be path or package name.  See
\code{\link{as.package}} for more information}
}
\description{
Use \code{Rmd} if you want a rich intermingling of code and data. Use
\code{md} for a basic README. \code{README.Rmd} will be automatically
added to \code{.Rbuildignore}.
}
\seealso{
Other infrastructure: \code{\link{infrastructure}},
  \code{\link{use_build_ignore}},
  \code{\link{use_data_raw}}, \code{\link{use_data}},
  \code{\link{use_news_md}}, \code{\link{use_package}}
}

