% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/infrastructure.R
\name{use_readme_rmd}
\alias{use_readme_rmd}
\title{Use README.Rmd}
\usage{
use_readme_rmd(pkg = ".")
}
\arguments{
\item{pkg}{package description, can be path or package name.  See
\code{\link{as.package}} for more information}

\item{hook}{Hook name. One of "pre-commit", "prepare-commit-msg",
"commit-msg", "post-commit", "applypatch-msg", "pre-applypatch",
"post-applypatch", "pre-rebase", "post-rewrite", "post-checkout",
"post-merge", "pre-push", "pre-auto-gc".}

\item{script}{Text of script to run}
}
\description{
This creates `README.Rmd` from template and adds to \code{.Rbuildignore}.
}
\seealso{
Other infrastructure: \code{\link{add_build_ignore}},
  \code{\link{use_build_ignore}};
  \code{\link{add_rstudio_project}},
  \code{\link{add_test_infrastructure}},
  \code{\link{add_test_infrastructure}},
  \code{\link{add_travis}}, \code{\link{add_travis}},
  \code{\link{add_travis}}, \code{\link{add_travis}},
  \code{\link{infrastructure}}, \code{\link{use_appveyor}},
  \code{\link{use_code_of_conduct}},
  \code{\link{use_coveralls}},
  \code{\link{use_cran_badge}},
  \code{\link{use_cran_comments}},
  \code{\link{use_package_doc}}, \code{\link{use_rcpp}},
  \code{\link{use_revdep}}, \code{\link{use_rstudio}},
  \code{\link{use_test}}, \code{\link{use_testthat}},
  \code{\link{use_travis}}, \code{\link{use_vignette}};
  \code{\link{use_data_raw}}; \code{\link{use_data}};
  \code{\link{use_package}}
}
\keyword{internal}

