% Generated by roxygen2 (4.0.2): do not edit by hand
\name{wd}
\alias{wd}
\title{Set working directory.}
\usage{
wd(pkg = ".", path = "")
}
\arguments{
\item{pkg}{package description, can be path or package name.  See
\code{\link{as.package}} for more information}

\item{path}{path within package. Leave empty to change working directory
to package directory.}
}
\description{
Set working directory.
}

