\name{wd}
\alias{wd}
\title{Set working directory.}
\usage{
wd(pkg = ".", path = NULL)
}
\arguments{
  \item{pkg}{package description, can be path or package
  name.  See \code{\link{as.package}} for more information}

  \item{path}{path within package. Leave empty to change
  working directory to package directory.}
}
\description{
  Set working directory.
}

